% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_dribble.R
\name{as_dribble}
\alias{as_dribble}
\alias{as_dribble.data.frame}
\alias{as_dribble.dribble}
\alias{as_dribble.character}
\alias{as_dribble.drive_id}
\alias{as_dribble.list}
\title{Coerce lists and character strings into Drive Files, \code{dribbles}.}
\usage{
as_dribble(x, ...)

\method{as_dribble}{data.frame}(x, ...)

\method{as_dribble}{dribble}(x, ...)

\method{as_dribble}{character}(x, ...)

\method{as_dribble}{drive_id}(x, ...)

\method{as_dribble}{list}(x, ...)
}
\arguments{
\item{x}{A list.}

\item{...}{Other arguments to pass on to individual methods.}
}
\description{
This is an S3 generic. dribble includes methods for
data frames and tibbles (adds \code{dribble} class), dribbles (returns
unchanged input), drive_ids, lists, and character vectors.
}
