% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_auth.R
\name{auth-config}
\alias{auth-config}
\alias{drive_api_key}
\alias{drive_oauth_app}
\title{Retrieve OAuth app or API key}
\usage{
drive_api_key()

drive_oauth_app()
}
\value{
An \code{\link[httr:oauth_app]{httr::oauth_app()}} or Google API key.
}
\description{
Retrieves the configured OAuth app and API key. Learn more in Google's
document \href{https://support.google.com/googleapi/answer/6158857?hl=en&ref_topic=7013279}{Credentials, access, security, and identity}.
By default, the app and API key are initialized to settings that ship with
googledrive. But the user can store their own app or key via
\code{\link[=drive_auth_config]{drive_auth_config()}}, i.e. overwrite the defaults.
}
\examples{
drive_api_key()
drive_oauth_app()

\dontrun{
drive_auth_config(api_key = "123")
drive_api_key()
}
}
\seealso{
Other auth functions: \code{\link{drive_auth_config}},
  \code{\link{drive_auth}}, \code{\link{drive_deauth}}
}
\concept{auth functions}
