% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{base_list_files}
\alias{base_list_files}
\title{Get the list of files from Googledrive}
\usage{
base_list_files(q = NULL, page_size = NULL, page_token = NULL,
  order_by = NULL, spaces = NULL, corpus = NULL)
}
\arguments{
\item{q}{Optional. A query for filtering the file results. See the "Search for Files" guide for supported syntax.}

\item{page_size}{Optional. The maximum number of files to return per page. Acceptable values are 1 to 1000,
inclusive. (Default: 100)}

\item{page_token}{Optional. The token for continuing a previous list request on the next page. This should be
set to the value of 'nextPageToken' from the previous response.}

\item{order_by}{Optional. A comma-separated list of sort keys. Valid keys are 'createdTime', 'folder',
'modifiedByMeTime', 'modifiedTime', 'name', 'quotaBytesUsed', 'recency', 'sharedWithMeTime',
'starred', and 'viewedByMeTime'. Each key sorts ascending by default, but may be reversed with the
'desc' modifier. Example usage: ?order_by=folder,modifiedTime desc,name.}

\item{spaces}{Optional. A comma-separated list of spaces to query within the corpus. Supported
values are 'drive', 'appDataFolder' and 'photos'.}

\item{corpus}{Optional. The source of files to list. Acceptable values are domain and user}
}
\description{
Utilizes the list_files Google Drive API endpoint to retrieves files and folder details.

This function provides the raw functionality of full access to the list_files API endpoint but is not
the recommended way of accessing the API. It would be better to use the higher level functions provided
within this package to access the files/folders/data that you need.
}

