% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_request.R
\name{build_request}
\alias{build_request}
\title{Build a request for the Google Drive v3 API}
\usage{
build_request(endpoint = character(), params = list(),
  token = drive_token(), .api_key = NULL)
}
\arguments{
\item{endpoint}{Character. Nickname for one of the documented Drive v3 API
endpoints. \emph{to do: list or link, once I've auto-generated those docs}}

\item{params}{Named list. Parameters destined for endpoint URL substitution
or, otherwise, the query.}

\item{token}{Drive token, obtained from \code{\link[=drive_auth]{drive_auth()}}}

\item{.api_key}{NULL for now.}
}
\value{
\code{list()}\cr Components are \code{method}, \code{path}, \code{query}, and \code{url},
suitable as input for \code{\link[=make_request]{make_request()}}. The \code{path} is post-substitution
and the \code{query} is a named list of all the input \code{params} that were not
used during this substitution. \code{url} is the full URL after prepending the
base URL for the Drive v3 API and appending an API key to the query.
}
\description{
Build a request, using some knowledge of the \href{https://developers.google.com/drive/v3/web/about-sdk}{Drive v3API}. Most users
should, instead, use higher-level wrappers that facilitate common tasks, such
as uploading or downloading Drive files. The functions here are intended for
internal use and for programming around the Drive API.
}
\details{
\itemize{
\item \code{build_request()} takes a nickname for an endpoint and uses the API spec to
look up the \code{path} and \code{method}. The \code{params} are checked for validity and
completeness with respect to the endpoint.
}
}
\examples{
\dontrun{
req <- build_request(
  "drive.files.get",
  list(
    fileId = "abc",
  )
)
req
}
}
