% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{delete_file}
\alias{delete_file}
\title{Delete a file in Google Drive}
\usage{
delete_file(fileID)
}
\arguments{
\item{fileID}{ID of the file in Google Drive}
}
\description{
Permanently deletes a file owned by the user without moving it to the trash.
}

