% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_auth.R
\name{drive_api_key}
\alias{drive_api_key}
\title{Retrieve built-in API key}
\usage{
drive_api_key()
}
\value{
A Google API key built into googledrive.
}
\description{
Retrieve built-in API key
}
\examples{
drive_api_key()
}
