% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_browse.R
\name{drive_browse}
\alias{drive_browse}
\title{Open Google Drive file in browser to edit}
\usage{
drive_browse(file)
}
\arguments{
\item{file}{\code{dribble} for the file you would like to edit}
}
\value{
\code{dribble}
}
\description{
Open Google Drive file in browser to edit
}
