% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_publish.R
\name{drive_check_publish}
\alias{drive_check_publish}
\title{Check if Google Drive file is published}
\usage{
drive_check_publish(file = NULL, verbose = TRUE)
}
\arguments{
\item{file}{\code{gfile} object representing the file you would like to check the published status of}

\item{verbose}{logical, indicating whether to print informative messages (default \code{TRUE})}
}
\value{
\code{gfile} object, a list with published information as a \code{tibble} under the list element \code{publish}
}
\description{
Check if Google Drive file is published
}
