% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_cp.R
\name{drive_copy}
\alias{drive_copy}
\title{Copy a Google Drive file.}
\usage{
drive_copy(file = NULL, verbose = TRUE)
}
\arguments{
\item{file}{Drive File, something that identifies the file of interest
on your Google Drive. This can be name(s)/path(s), marked file id(s),
or a \code{dribble}.}

\item{verbose}{Logical, indicating whether to print informative messages
(default \code{TRUE}).}
}
\value{
An object of class \code{\link{dribble}}, a tibble with
   one row per file.
}
\description{
Copy a Google Drive file.
}
\examples{
## this will make a copy of the file "chickwts.rda"
## in the same folder as the original file named
## "Copy of chickwts.rda".
drive_copy("chickwts.rda")
}
