% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_corpus.R
\name{drive_corpus}
\alias{drive_corpus}
\title{Access Team Drive files}
\usage{
drive_corpus(teamDriveId = NULL, corpora = NULL,
  includeTeamDriveItems = NULL)
}
\arguments{
\item{teamDriveId}{Character, a Team Drive id.}

\item{corpora}{Character, specifying the search collection. Only relevant in
the Team Drives context.}

\item{includeTeamDriveItems}{Logical, googledrive always sets this to \code{TRUE}.}
}
\description{
The Drive API requires extra information in order to look for
files on a Team Drive. \code{\link[=drive_find]{drive_find()}} and \code{\link[=drive_get]{drive_get()}} accept this extra
information via \code{team_drive} and/or \code{corpora}. Regard these two functions
as the official "port of entry" for Team Drive files and folders. Use them
to capture your target(s) in a \code{\link{dribble}} to pass along to other
googledrive functions. The general flexibility to refer to files by name,
path, or id does not apply to Team Drive files. While it's always a good
idea to get things into a \code{\link{dribble}} early, for files on a Team Drive,
it's a requirement.

\code{team_drive} and/or \code{corpora} are pre-processed by \code{drive_corpus()} to obtain
query parameters.

Note:
  \href{https://gsuite.google.com/learning-center/products/drive/get-started-team-drive/}{Team Drives}
  are only available to users of certain enhanced Google services, such as G
  Suite Enterprise, G Suite Business, or G Suite for Education.

If you want to search one specific Team Drive, pass its name,
id, or \code{\link{dribble}} to \code{team_drive} somewhere in the call, like so:\preformatted{drive_find(..., team_drive = "i-am-a-teamdrive-name")
drive_find(..., team_drive = as_id("i-am-a-teamdrive-id"))
drive_find(..., team_drive = teamdrive_dribble)
}

The value of \code{team_drive} is pre-processed with \code{\link[=as_teamdrive]{as_teamdrive()}}.

To search other collections, pass the \code{corpora} parameter somewhere in the
call, like so:\preformatted{drive_find(..., corpora = "user,allTeamDrives")
}

Possible values of \code{corpora} and what they mean:
\itemize{
\item \code{"user"}: Queries files that the user has accessed, including both Team
and non-Team Drive files.
\item \code{"user,allTeamDrives"}: Queries files that the user has accessed and all
Team Drives in which they are a member. Note that both \code{"user"} and
\code{"teamDrive"} are preferable to \code{"allTeamDrives"}, in terms of efficiency.
\item \code{"domain"}: Queries files that are shared to the domain, including both
Team Drive and non-Team Drive files.
\item \code{"teamDrive"}: Queries all items in one specific Team Drive.
\code{teamDriveId} must be also specified elsewhere in the request. The
googledrive workflow in this case is to use the \code{team_drive} argument as
described above and allow googledrive to infer the need to send
\code{corpora = "teamDrive"}.
}
}
\details{
Team Drive support:
googledrive implements Team Drive support as outlined here:
\itemize{
\item \url{https://developers.google.com/drive/v3/web/enable-teamdrives#including_team_drive_content_fileslist}
}

Users shouldn't need to know any of this, but here are details for the
curious. The extra information needed to search Team Drives consists of:
\itemize{
\item \code{corpora}: Where to search? Described above.
\item \code{teamDriveId}: The id of a specific Team Drive. Only allowed -- and also
absolutely required -- when \code{corpora = "teamDrive"}. When user specifies
a Team Drive, googledrive sends its id and also infers that \code{corpora}
should be set to \code{"teamDrive"} and sent.
\item \code{includeTeamDriveItems}: Do you want to see Team Drive items? Obviously,
this should be \code{TRUE} and googledrive sends this whenever Team Drive
parameters are detected
\item \code{supportsTeamDrives}: Does the sending application (googledrive, in this
case) know about Team Drives? Obviously, this should be \code{TRUE} and
googledrive sends it for all applicable endpoints, all the time.
}
}
\examples{
drive_corpus(teamDriveId = "123456789")
drive_corpus(corpora = "user")
drive_corpus(corpora = "user,allTeamDrives")
drive_corpus(corpora = "domain")

\dontrun{
## this will error because `corpora = "teamDrive"` also requires the id
drive_corpus(corpora = "teamDrive")

## this will error because `corpora = "domain"` forbids inclusion of id
drive_corpus(corpora = "domain", teamDriveId = "123456789")
}
}
\seealso{
\code{\link[=drive_find]{drive_find()}}
}
