% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_delete.R
\name{drive_delete}
\alias{drive_delete}
\title{Delete file from Google Drive}
\usage{
drive_delete(file = NULL, verbose = TRUE)
}
\arguments{
\item{file}{\code{dribble} representing the file you would like to
delete}

\item{verbose}{logical, indicating whether to print informative messages
(default \code{TRUE})}
}
\value{
logical, indicating whether the delete succeeded
}
\description{
Delete file from Google Drive
}
