% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_download.R
\name{drive_download}
\alias{drive_download}
\title{Download a file from Google Drive}
\usage{
drive_download(file = NULL, to = NULL, overwrite = FALSE,
  fun = writeBin, verbose = TRUE)
}
\arguments{
\item{file}{Drive File, something that identifies the file of interest
on your Google Drive. This can be name(s)/path(s), marked file id(s),
or a \code{dribble}.}

\item{to}{Character, local path for the file download. Will default to
saving to the working directory by its name on Google Drive.}

\item{overwrite}{A logical scalar, do you want to overwrite a local file
if such exists?}

\item{fun}{Function used to write the downloaded file type.}

\item{verbose}{Logical, indicating whether to print informative messages
(default \code{TRUE}).}
}
\description{
Download a file from Google Drive
}
