% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_example.R
\name{drive_example}
\alias{drive_example}
\title{Get path to example file}
\usage{
drive_example(path = NULL)
}
\arguments{
\item{path}{Name of file. If \code{NULL}, the example files will be listed.}
}
\description{
googledrive comes bundled with a few small files to use in examples. This
function make them easy to access.
}
\examples{
drive_example()
drive_example("chicken.jpg")
}
