% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_download.R
\name{drive_extract_file}
\alias{drive_extract_file}
\title{Extract raw file from \code{dribble}.}
\usage{
drive_extract_file(x)
}
\arguments{
\item{x}{A \code{\link{dribble}} with one row and a \code{raw_file} column, obtained after
running \code{\link[=drive_download]{drive_download()}}.}
}
\description{
Extract raw file from \code{dribble}.
}
\examples{
\dontrun{
## download a .csv file
drive_download("chickwts.csv") \%>\%
  drive_extract_file() \%>\%
  write.csv("chickwts.csv")

## download an .rda file
drive_download("chickwts.rda") \%>\%
  drive_extract_file() \%>\%
  writeBin("chickwts.rda")
}

}
