% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_file.R
\name{drive_file}
\alias{drive_file}
\title{Create a Google Drive file object}
\usage{
drive_file(id, ...)
}
\arguments{
\item{id}{character, Google drive id for file of interest}

\item{...}{name-value pairs to query the API}
}
\value{
object of class \code{gfile} and \code{list}
}
\description{
Create a Google Drive file object
}
