% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_get.R
\name{drive_get}
\alias{drive_get}
\title{Get Drive files by id}
\usage{
drive_get(id)
}
\arguments{
\item{id}{Character, a Drive file id, such as you might see in the URL when
visiting a file on Google Drive.}
}
\value{
dribble-return
}
\description{
Warning: the name of this function is likely to change.
}
\examples{
\dontrun{
drive_get("abcdefgeh12345678")
}
}
