% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_auth.R
\name{drive_has_token}
\alias{drive_has_token}
\title{Is there a token on hand?}
\usage{
drive_has_token()
}
\value{
Logical.
}
\description{
Reports whether googledrive has stored a token, ready for use in downstream
requests.
}
\examples{
drive_has_token()
}
\seealso{
Other low-level API functions:
\code{\link[=drive_token]{drive_token()}},
\code{\link[=request_generate]{request_generate()}},
\code{\link[=request_make]{request_make()}}
}
\concept{low-level API functions}
