% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_publish.R
\name{drive_is_published}
\alias{drive_is_published}
\title{Check if Google Drive file is published}
\usage{
drive_is_published(file = NULL, verbose = TRUE)
}
\arguments{
\item{file}{\code{dribble} or \code{drive_id} object representing the file you would like to check the published status of}

\item{verbose}{logical, indicating whether to print informative messages (default \code{TRUE})}
}
\value{
\code{dribble} object
}
\description{
Check if Google Drive file is published
}
