% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_ls.R
\name{drive_list}
\alias{drive_list}
\title{List contents of a folder.}
\usage{
drive_list(path = "~/", pattern = NULL, type = NULL, ...)
}
\arguments{
\item{path}{Character. A single folder on Google Drive whose contents you
want to list. Can be an actual path (character), a file id marked with \code{\link[=as_id]{as_id()}}, or
a \code{\link{dribble}}.}

\item{pattern}{Character. If provided, only the files whose names match this
regular expression are returned.}

\item{type}{Character. If provided, only files of this type will be returned.
Can be anything that \code{\link[=drive_mime_type]{drive_mime_type()}} knows how to handle. This is
processed by googledrive and sent as a query parameter.}

\item{...}{Query parameters to pass along to the API query.}
}
\value{
An object of class \code{\link{dribble}}, a tibble with
   one row per file.
}
\description{
List contents of a folder.
}
\examples{
\dontrun{
## get contents of the folder 'abc' (non-recursive)
drive_list("abc")

## get contents of folder 'abc' that contain the
## letters 'def'
drive_list(path = "abc", pattern = "def")

## get all Google spreadsheets in folder 'abc'
## that contain the letters 'def'
drive_list(path = "abc", pattern = "def", type = "spreadsheet")
}
}
