% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_ls.R
\name{drive_ls}
\alias{drive_ls}
\title{List Drive files.}
\usage{
drive_ls(path = "~/", pattern = NULL, ...)
}
\arguments{
\item{path}{Character. A single path on Google Drive whose contents you
want to list.}

\item{pattern}{Character. If provided, only the files whose names match this
regular expression are returned.}

\item{...}{Parameters to pass along to the API query.}
}
\value{
An object of class \code{dribble}, a tibble with
   one row per file.
}
\description{
List Drive files.
}
\examples{
\dontrun{
## get contents of the folder 'abc' (non-recursive)
drive_ls(path = "abc")

## get contents of folder 'abc' that contain the
## letters 'def'
drive_ls(path = "abc", pattern = "def")
}
}
