% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_mv.R
\name{drive_move}
\alias{drive_move}
\title{Move Google Drive file.}
\usage{
drive_move(file = NULL, name = NULL, path = NULL, verbose = TRUE)
}
\arguments{
\item{file}{Drive File, something that identifies the file of interest
on your Google Drive. This can be name(s)/path(s), marked file id(s),
or a \code{dribble}.}

\item{name}{Character, name you would like the moved file to have. Will
default to its current name.}

\item{path}{Drive File, something that identifies where the file should live
on your Google Drive. Can be an actual path (character), a file id marked
with \code{\link{as_id}()}, or a \code{\link{dribble}}.}

\item{verbose}{Logical, indicating whether to print informative messages
(default \code{TRUE}).}
}
\value{
An object of class \code{\link{dribble}}, a tibble with
   one row per file.
}
\description{
Move Google Drive file.
}
