% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_permission.R
\name{drive_permission_list}
\alias{drive_permission_list}
\title{Work with Drive file permissions}
\usage{
drive_permission_list(file, ...)
}
\arguments{
\item{file}{Something that identifies the file(s) of interest on your Google
Drive. Can be a character vector of names/paths, a character vector of file
ids or URLs marked with \code{\link[=as_id]{as_id()}}, or a \code{\link{dribble}}.}
}
\value{
\itemize{
\item \code{drive_permission_list()}: A tibble with one row per permission, per file.
The \code{name} and \code{id} columns identify the associated file. The remaining
columns are determined dynamically from the permission data returned by
the Drive API. This should always include \code{permission_id}, as well as the
permission \code{type} and \code{role}.
}
}
\description{
A family of functions for working with Drive file permissions.
\itemize{
\item \code{drive_permission_list()}: Lists the permissions associated with each
input file. Useful for getting a \code{permission_id}, if you need to
delete a permission \emph{not implemented yet}.
\item \emph{More functions will be added, especially, \code{drive_permission_delete()}.}
\item \code{drive_share()} has been in googledrive for a long time and, morally, it's
\verb{drive_permission_create().} \emph{We will probably add
\code{drive_permission_create()} and make \code{drive_share()} an alias.}
}
}
\examples{
\dontshow{if (drive_has_token()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Create a file to share
file <- drive_example_remote("chicken_doc") \%>\%
  drive_cp(name = "chicken-permissions.txt")

# View the initial permissions
drive_permission_list(file)

# Let a specific person comment
file <- file \%>\%
  drive_share(
    role = "commenter",
    type = "user",
    emailAddress = "susan@example.com"
  )

# Allow "anyone with the link" to read the file
drive_share_anyone(file)

# View the permissions again
drive_permission_list(file)

# Clean up
drive_rm(file)
\dontshow{\}) # examplesIf}
}
\seealso{
Overview of the \code{permissions} API:
\itemize{
\item \url{https://developers.google.com/drive/api/v3/reference/permissions}
}

Drive roles and permissions are described here:
\itemize{
\item \url{https://developers.google.com/drive/api/v3/ref-roles}
}
}
