% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_publish.R
\name{drive_publish}
\alias{drive_publish}
\title{Publish Google Drive file}
\usage{
drive_publish(file, ..., verbose = TRUE)
}
\arguments{
\item{file}{Something that identifies the file(s) of interest on your Google
Drive. Can be a character vector of names/paths, a character vector of file
ids or URLs marked with \code{\link{as_id}()}, or a \code{\link{dribble}}.}

\item{...}{Name-value pairs to add to the API request body, for example
\code{publishAuto = FALSE} will ensure that each subsequent revision will not be
automatically published (default here is \code{publishAuto = TRUE} and \code{publishedOutsideDomain = TRUE}).}

\item{verbose}{Logical, indicating whether to print informative messages
(default \code{TRUE}).}
}
\value{
An object of class \code{\link{dribble}}, a tibble with one row per
  item.
}
\description{
Publish a Google Drive file to the web. You can only publish "Google" files, such
as Google Documents, Google Spreadsheets, or Google Slides.
}
\examples{
\dontrun{
## Upload file to publish
file <- drive_upload(R.home('doc/BioC_mirrors.csv'),
  type = "spreadsheet")

## Publish file
file <- drive_publish(file)

## Clean up
drive_rm(file)
}
}
\seealso{
Wraps the \code{revisions.update} endpoint:
\itemize{
\item \url{https://developers.google.com/drive/v3/reference/revisions/update}
}
}
