% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_reveal_path.R
\name{drive_reveal_path}
\alias{drive_reveal_path}
\title{Get paths to Drive files.}
\usage{
drive_reveal_path(file)
}
\arguments{
\item{file}{Something that identifies the file(s) of interest on your Google
Drive. Can be a character vector of names/paths, a character vector of file
ids or URLs marked with \code{\link{as_id}()}, or a \code{\link{dribble}}.}
}
\value{
An object of class \code{\link{dribble}}, a tibble with one row per
  item.
}
\description{
Even though you know a Drive file's name or id, you don't
necessarily know its path(s). This function takes Drive files, specified
via a \code{\link{dribble}} or a vector of file ids or URLs, and returns a dribble
that includes a \code{path} variable.

Note that Google Drive does NOT behave like your local file
  system:
  \itemize{
  \item File and folder names need not be unique, even at a given level of
  the hierarchy. A single name or file path can be associated with multiple
  files (or zero or exactly one).
  \item A file can have more than one direct parent. This implies that a
  single file can be represented by multiple paths.
  }

Bottom line: Do not assume there is a one-to-one relationship
  between file name or path and a Drive file or folder. A file id is unique,
  which is why googledrive workflows favor storage in a
  \code{\link{dribble}}, which holds metadata aimed at computers (file id)
  and humans (file name). Finally, note also that a folder is just a specific
  type of file on Drive.
}
\examples{
\dontrun{
## 'root' is a special file id that always represents your root folder
drive_get(id = "root") \%>\% drive_reveal("path")
}
}
