% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_auth.R
\name{drive_set_token}
\alias{drive_set_token}
\title{API to set OAuth token for googledrive
This API does not refresh the provided token and use it as is.}
\usage{
drive_set_token(token)
}
\arguments{
\item{token}{\itemize{
\item OAuth token for the googledrive
}}
}
\description{
API to set OAuth token for googledrive
This API does not refresh the provided token and use it as is.
}
\seealso{
Other auth functions: \code{\link{drive_auth_configure}},
  \code{\link{drive_auth}}, \code{\link{drive_deauth}}
}
\concept{auth functions}
