% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_share.R
\name{drive_share}
\alias{drive_share}
\title{Update Google Drive file share permissions}
\usage{
drive_share(file = NULL, role = NULL, type = NULL, ..., verbose = TRUE)
}
\arguments{
\item{file}{\code{dribble} or \code{drive_id} object representing the file you
would like to share}

\item{role}{The role granted by this permission. Valid values are:
\itemize{
\item organizer
\item owner
\item writer
\item commenter
\item reader
}}

\item{type}{The type of the grantee. Valid values are:
\itemize{
\item user
\item group
\item domain
\item anyone
}}

\item{...}{name-value pairs to add to the API request body}

\item{verbose}{logical, indicating whether to print informative messages
(default \code{TRUE})}
}
\value{
\code{dribble} object updated with new sharing information
}
\description{
Update Google Drive file share permissions
}
