% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_show_path.R
\name{drive_show_path}
\alias{drive_show_path}
\title{Get paths to Drive files.}
\usage{
drive_show_path(file)
}
\arguments{
\item{file}{Something that identifies the file(s) of interest on your Google
Drive. Can be a character vector of names/paths, a character vector of file
ids or URLs marked with \code{\link{as_id}()}, or a \code{\link{dribble}}.}
}
\value{
An object of class \code{\link{dribble}}, a tibble with one row per
  item.
}
\description{
Even though you know a Drive file's name or id, you don't
necessarily know its path(s). This function takes Drive files, specified
via a \code{\link{dribble}} or a vector of file ids or URLs, and returns a dribble
that includes a \code{path} variable. Note that Google Drive does NOT behave
like your local file system:
\itemize{
\item A single Drive file can potentially be represented by multiple paths,
since a file or folder can have multiple direct parents. So \code{n} inputs
could result in more than \code{n} outputs.
\item Multiple Drive files could be associated with the same path, because
file and folder names need not be unique, even at a given level of the
hierarchy. So \code{n} inputs could result in fewer than \code{n} unique values in
the \code{path} variable.
}
}
\examples{
\dontrun{
## 'root' is a special file id that always represents your root folder
drive_get(id = "root") \%>\% drive_show_path()
}
}
