% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_publish.R
\name{drive_show_publish}
\alias{drive_show_publish}
\title{Add columns with publication information to your dribble}
\usage{
drive_show_publish(file, verbose = TRUE)
}
\arguments{
\item{file}{Something that identifies the file(s) of interest on your Google
Drive. Can be a character vector of names/paths, a character vector of file
ids or URLs marked with \code{\link{as_id}()}, or a \code{\link{dribble}}.}

\item{verbose}{Logical, indicating whether to print informative messages
(default \code{TRUE}).}
}
\value{
An object of class \code{\link{dribble}}, a tibble with one row per
  item.
}
\description{
Add columns with publication information to your dribble
}
\examples{
\dontrun{
## Upload file to check publication status
file <- drive_upload(R.home('doc/BioC_mirrors.csv'),
  type = "spreadsheet")

## Check publication status
drive_show_publish(file)

## Publish file
drive_publish(file)

## Check publication status agian
drive_show_publish(file)

## Unpublish file
drive_unpublish(file)

## Clean up
drive_rm(file)
}
}
