% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_share.R
\name{drive_show_sharing}
\alias{drive_show_sharing}
\title{Add sharing column(s) to your dribble}
\usage{
drive_show_sharing(file, display = "name", role = "owner")
}
\arguments{
\item{file}{Something that identifies the file(s) of interest on your Google
Drive. Can be a character vector of names/paths, a character vector of file
ids or URLs marked with \code{\link{as_id}()}, or a \code{\link{dribble}}.}

\item{display}{Character. The value you'd like displayed for who has share permissions.
Defaults to "name". Valid values are:
\itemize{
\item id
\item name
\item type
\item email
}}

\item{role}{Character. The role(s) you'd like to see the permissions for. A column with
permission information based on the \code{display} value chosen will be added for each role.
Defaults to "owner". Valid values are any combination of:
\itemize{
\item organizer
\item owner
\item writer
\item commenter
\item reader
}}
}
\value{
An object of class \code{\link{dribble}}, a tibble with one row per
  item.
}
\description{
Add sharing column(s) to your dribble
}
\examples{
\dontrun{
## Upload a file to view sharing permissions
file <- drive_upload(
   system.file("DESCRIPTION"),
   type = "document"
   )

## Add default sharing information (the name of the owner)
drive_show_sharing(file)

## Add sharing information (name) for those with role "owner" and "commenter"
drive_show_sharing(file, role = c("owner", "commenter"))

## Add sharing information (email address) for all "readers"
drive_show_sharing(file, display = "email", role = "reader")

## Clean up
drive_rm(file)
}
}
