% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_trash.R
\name{drive_show_trash}
\alias{drive_show_trash}
\title{Get files in Drive Trash.}
\usage{
drive_show_trash()
}
\value{
An object of class \code{\link{dribble}}, a tibble with one row per
  item.
}
\description{
Get files in Drive Trash.
}
