% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_trash.R
\name{drive_trash}
\alias{drive_trash}
\title{Trash file on Google Drive.}
\usage{
drive_trash(file = NULL, verbose = TRUE)
}
\arguments{
\item{file}{Drive File, something that identifies the file of interest
on your Google Drive. This can be name(s)/path(s), marked file id(s),
or a \code{dribble}.}

\item{verbose}{Logical, indicating whether to print informative messages
(default \code{TRUE}).}
}
\value{
Logical, indicating whether the trashing succeeded.
}
\description{
Trash file on Google Drive.
}
\examples{
\dontrun{
## Trash a single file in the trash.
drive_trash("chickwts.csv")

## Trash multiple files in trash.
drive_trash(c("abc", "def"))
}
}
