% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_publish.R
\name{drive_unpublish}
\alias{drive_unpublish}
\title{Unpublish Google Drive file}
\usage{
drive_unpublish(file = NULL, ..., verbose = TRUE)
}
\arguments{
\item{file}{Drive File, something that identifies the file of interest
on your Google Drive. This can be name(s)/path(s), marked file id(s),
or a `dribble`.}

\item{...}{name-value pairs to add to the API request body.}

\item{verbose}{logical, indicating whether to print informative messages
(default `TRUE`)}
}
\value{
Drive File. An object of class `dribble`, a tibble with
   one row per file.
}
\description{
Unpublish Google Drive file
}
