% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_trash.R
\name{drive_untrash}
\alias{drive_untrash}
\title{Untrash file on Google Drive.}
\usage{
drive_untrash(file = NULL, verbose = TRUE)
}
\arguments{
\item{file}{Drive File, something that identifies the file of interest
on your Google Drive. This can be name(s)/path(s), marked file id(s),
or a \code{dribble}.}

\item{verbose}{Logical, indicating whether to print informative messages
(default \code{TRUE}).}
}
\value{
Logical, indicating whether the untrashing succeeded.
}
\description{
Untrash file on Google Drive.
}
\examples{
\dontrun{
## Untrash a single file in the trash.
drive_untrash("chickwts.csv")

## Untrash multiple files in trash.
drive_untrash(c("abc", "def"))
}
}
