% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_user.R
\name{drive_user}
\alias{drive_user}
\title{Information on the user and user's Drive.}
\usage{
drive_user(fields = "user", ..., verbose = TRUE)
}
\arguments{
\item{fields}{The fields the user would like output - by default only \code{user},
which will display as detailed above.}

\item{...}{Name-value pairs to query the API.}

\item{verbose}{Logical, indicating whether to print informative messages
(default \code{TRUE}).}
}
\value{
A list of class \code{drive_user} with user's data.
}
\description{
Reveals information about the user associated with the current token.
}
\seealso{
Wraps the \code{about.get} endpoint:
\itemize{
\item \url{https://developers.google.com/drive/v3/reference/about/get}
}
}
