% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_trash.R
\name{drive_view_trash}
\alias{drive_view_trash}
\title{View files in Google Drive Trash.}
\usage{
drive_view_trash()
}
\value{
An object of class \code{\link{dribble}}, a tibble with
   one row per file.
}
\description{
View files in Google Drive Trash.
}
