% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_download.R
\name{drive_write_csv}
\alias{drive_write_csv}
\title{Write Google File as a csv}
\usage{
drive_write_csv(file = NULL, path = NULL, verbose = TRUE)
}
\arguments{
\item{file}{Drive File, something that identifies the file of interest
on your Google Drive. This can be name(s)/path(s), marked file id(s),
or a \code{dribble}.}

\item{path}{Character. Path to write to.}

\item{verbose}{Logical, indicating whether to print informative messages
(default \code{TRUE}).}
}
\description{
Write Google File as a csv
}
