% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{expose}
\alias{expose}
\title{An expose object}
\usage{
expose()
}
\description{
\code{expose()} returns a sentinel object, similar in spirit to \code{NULL}, that tells
the calling function to return its internal data structure. googledrive
stores alot of information about the Drive API, MIME types, etc., internally
and then exploits it in helper functions, like \code{\link[=drive_mime_type]{drive_mime_type()}},
\code{\link[=drive_fields]{drive_fields()}}, \code{\link[=drive_endpoints]{drive_endpoints()}}, etc. We use these objects to
provide nice defaults, check input validity, or lookup something cryptic,
like MIME type, based on something friendlier, like a file extension. Pass
\code{expose()} to such a function if you want to inspect its internal object, in
its full glory. This is inspired by the \code{waiver()} object in ggplot2.
}
\examples{
drive_mime_type(expose())
drive_fields(expose())
}
\keyword{internal}
