% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_browse.R
\name{gd_browse}
\alias{gd_browse}
\title{Open Google Drive file in browser to edit}
\usage{
gd_browse(file)
}
\arguments{
\item{file}{\code{drive_file} object for the file you would like to edit}
}
\description{
Open Google Drive file in browser to edit
}
