% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_auth.R
\name{gd_deauth}
\alias{gd_deauth}
\title{Suspend authorization}
\usage{
gd_deauth(clear_cache = TRUE, verbose = TRUE)
}
\arguments{
\item{clear_cache}{logical indicating whether to disable the
\code{.httr-oauth} file in working directory, if such exists, by renaming
to \code{.httr-oauth-SUSPENDED}}
}
\description{
Suspend \code{\link{googledrive}}' authorization to place requests to the
Drive and Sheets APIs on behalf of the authenticated user.
}
\examples{
\dontrun{
gd_deauth()
}
}
\seealso{
Other auth.functions: \code{\link{gd_auth}}
}

