% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_delete.R
\name{gd_delete}
\alias{gd_delete}
\title{Delete file from Google Drive}
\usage{
gd_delete(file = NULL, verbose = TRUE)
}
\arguments{
\item{file}{\code{drive_file} object representing the file you would like to
delete}

\item{verbose}{logical, indicating whether to print informative messages
(default \code{TRUE})}
}
\value{
logical, indicating whether the delete succeeded
}
\description{
Delete file from Google Drive
}
