% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_file.R
\name{gd_file}
\alias{gd_file}
\title{Create a Google Drive file object}
\usage{
gd_file(id, ...)
}
\arguments{
\item{id}{character, Google drive id for file of interest}

\item{...}{name-value pairs to query the API}
}
\value{
object of class \code{drive_file} and \code{list}
}
\description{
Create a Google Drive file object
}

