% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_get.R
\name{gd_get}
\alias{gd_get}
\title{Google Drive Get}
\usage{
gd_get(id, ..., fields = default_fields, auth = TRUE)
}
\arguments{
\item{id}{document id obtained from `gd_ls`}

\item{...}{name-value pairs to query the API}

\item{fields}{user can input an vector of fields - defaults to all fields}

\item{auth}{default TRUE}
}
\value{
tibble
}
\description{
Google Drive Get
}

