% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_get_id.R
\name{gd_get_id}
\alias{gd_get_id}
\title{Get Google Drive id}
\usage{
gd_get_id(search, ..., fixed = FALSE, verbose = TRUE)
}
\arguments{
\item{search}{character, regular expression to determin document title to output the Google Drive id
If it is \code{NULL} (default), the most recently modified document id will be output.}

\item{...}{name-value pairs to query the API}

\item{fixed}{logical, from \code{\link{grep}}. If \code{TRUE}, \code{search} is exactly matched to a
document's name on Google Drive.}

\item{verbose}{logical, indicating whether to print informative messages (default \code{TRUE})}
}
\value{
character, Google Drive id
}
\description{
Get Google Drive id
}

