% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_mv.R
\name{gd_mv}
\alias{gd_mv}
\title{Move Google Drive file}
\usage{
gd_mv(file = NULL, folder = NULL)
}
\arguments{
\item{file}{\code{drive_file} object for the file you would like to move}

\item{folder}{\code{drive_file} object for the folder you would like to move the file to}
}
\value{
\code{drive_file} that was moved
}
\description{
Move Google Drive file
}
