% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_open.R
\name{gd_open}
\alias{gd_open}
\title{Open Google Drive file in browser to edit}
\usage{
gd_open(file)
}
\arguments{
\item{file}{\code{drive_file} object for the file you would like to edit}
}
\description{
Open Google Drive file in browser to edit
}
