% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_publish.R
\name{gd_publish}
\alias{gd_publish}
\title{Publish Google Drive file}
\usage{
gd_publish(file, verbose = TRUE)
}
\arguments{
\item{file}{\code{drive_file} object representing the file you would like to publish}

\item{verbose}{logical, indicating whether to print informative messages (default \code{TRUE})}
}
\value{
logical, whether the file was successfully published
}
\description{
Publish Google Drive file
}
