% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_share.R
\name{gd_share}
\alias{gd_share}
\title{Update Google Drive file share permissions}
\usage{
gd_share(file, role = NULL, type = NULL, email = NULL, message = NULL,
  ..., verbose = TRUE)
}
\arguments{
\item{file}{\code{drive_file} object representing the file you would like to delete}

\item{role}{The role granted by this permission. Valid values are:
\itemize{
\item organizer
\item owner
\item writer
\item commenter
\item reader
}}

\item{type}{The type of the grantee. Valid values are:
\itemize{
\item user
\item group
\item domain
\item anyone
}}

\item{email}{The email address of the user or group to which this permission refers.}

\item{message}{A custom message to include in the notification email.}

\item{...}{name-value pairs to add to the API request body}

\item{verbose}{logical, indicating whether to print informative messages (default \code{TRUE})}
}
\value{
logical, indicating whether the permissions update succeeded
}
\description{
Update Google Drive file share permissions
}
