% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_share_link.R
\name{gd_share_link}
\alias{gd_share_link}
\title{Retrieves Google Drive file's share link}
\usage{
gd_share_link(file)
}
\arguments{
\item{file}{\code{drive_file} object representing the file you would like to
retrieve the link for}
}
\value{
character, link to Google Drive file
}
\description{
Retrieves Google Drive file's share link
}
