% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_upload.R
\name{gd_upload}
\alias{gd_upload}
\title{Upload a file to Google Drive}
\usage{
gd_upload(file = NULL, name = NULL, overwrite = FALSE, type = NULL, ...,
  verbose = TRUE)
}
\arguments{
\item{file}{character, path the the file you'd like to upload}

\item{name}{character, what you'd like the uploaded file to be called on Google Drive}

\item{overwrite}{logical, do you want to overwrite file already on Google Drive}

\item{type}{if type = \code{NULL}, will force type as follows:
\itemize{
\item \strong{document}: .doc, .docx, .txt, .rtf., .html, .odt, .pdf, .jpeg, .png, .gif,.bmp
\item \strong{spreadsheet}: .xls, .xlsx, .csv, .tsv, .tab, .xlsm, .xlt, .xltx, .xltm,
.ods
\item \strong{presentation}: .opt, .ppt, .pptx, .pptm
otherwise you can specify \code{document}, \code{spreadsheet}, or \code{presentation}. Files with no extension will be assumed to be a \code{folder}
}}

\item{...}{name-value pairs to query the API}

\item{verbose}{logical, indicating whether to print informative messages (default \code{TRUE})}
}
\value{
object of class \code{drive_file} and \code{list} that contains uploaded file's information
}
\description{
Upload a file to Google Drive
}
