% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_user.R
\name{gd_user}
\alias{gd_user}
\title{Pull information about the Google Drive user}
\usage{
gd_user(fields = "user", ..., verbose = TRUE)
}
\arguments{
\item{fields}{the fields the user would like output - by default only `user`, which will display as detailed above.}

\item{...}{name-value pairs to query the API}

\item{verbose}{Logical, indicating whether to print informative messages (default \code{TRUE})}
}
\value{
a list of class \code{drive_user} with user's data
}
\description{
Pull information about the Google Drive user
}

