% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googledrive.R
\docType{package}
\name{googledrive}
\alias{googledrive}
\alias{googledrive-package}
\title{googledrive: Controlling Googledrive with R}
\description{
This package aims is to give R users the capability to have full programmatic
access to Google Drive files.
}

