% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_user.R
\name{guser}
\alias{guser}
\title{Google Drive User Information}
\usage{
guser(fields = "user", ...)
}
\arguments{
\item{fields}{fields to query, default is \code{user}.}

\item{...}{name-value pairs to query the API}
}
\value{
list of class \code{guser} with user's information
}
\description{
Google Drive User Information
}
\keyword{internal}
