% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dribble.R
\name{is_one}
\alias{is_one}
\title{Checks whether dribble contains exactly one Drive file.}
\usage{
is_one(file, .what = "file")
}
\arguments{
\item{file}{Drive File, something that identifies the file of interest
on your Google Drive. This can be name(s)/path(s), marked file id(s),
or a \code{dribble}.}

\item{.what}{Character, description of intput for informative
error message.}
}
\description{
This will return the input \code{dribble} if it contains exactly
one Drive file, and will error otherwise.
}
