% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{list_files}
\alias{list_files}
\title{Get the list of files from Googledrive}
\usage{
list_files(q = NULL, pageSize = NULL, pageToken = NULL, orderBy = NULL,
  spaces = NULL, corpus = NULL)
}
\description{
Get the list of files from Googledrive
}

