% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{move_file}
\alias{move_file}
\title{Add or remove file from folders}
\usage{
move_file(file_id, add_folders = NULL, remove_folders = NULL)
}
\arguments{
\item{file_id}{The ID of the file.}

\item{addFolders}{A character vector of folder Ids}

\item{removeFolders}{A character vector of folder Ids}
}
\description{
Allows you to move file around. Due to the nature of Google drive which allows multiple
files with the same names to coexist in the same folder, we can technically "hook" files into the
folders. At the same time, there is a convenient feature of being able to "hook" a file into
multiple folders at the same time which basically means a file can be in 2 folders at once.
}

