% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_response.R
\name{process_response}
\alias{process_response}
\title{Process a response from the Google Drive v3 API}
\usage{
process_response(res)
}
\arguments{
\item{res}{Object of class \code{response} from \link{httr}.}
}
\value{
List.
}
\description{
Process a response from the Google Drive v3 API
}
\seealso{
Other low-level API functions: \code{\link{drive_token}},
  \code{\link{generate_request}},
  \code{\link{make_request}}
}
