% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_drive_create.R
\name{team_drive_create}
\alias{team_drive_create}
\title{Create a new Team Drive}
\usage{
team_drive_create(name, verbose = TRUE)
}
\arguments{
\item{name}{Character. Name of the new Team Drive. Must be non-empty and not
entirely whitespace.}

\item{verbose}{Logical, indicating whether to print informative messages
(default \code{TRUE}).}
}
\value{
An object of class \code{\link{dribble}}, a tibble with one row per item.
}
\description{
Team Drives have been rebranded as \emph{shared drives} and, as of googledrive
v2.0.0, all \verb{team_drive_*()} functions have been deprecated, in favor of
\verb{shared_drive_*()} successors.

The changes in googledrive reflect that the Team Drives resource collection
has been deprecated in the Drive API v3, in favor of the new (shared) Drives
resource collection. Read more
\itemize{
\item \url{https://cloud.google.com/blog/products/application-development/upcoming-changes-to-the-google-drive-api-and-google-picker-api}
}
}
\examples{
\dontrun{
team_drive_create("my-awesome-team-drive")

## clean up
team_drive_rm("my-awesome-team-drive")
}
}
\seealso{
Wraps the \code{teamdrives.create} endpoint:
\itemize{
\item \url{https://developers.google.com/drive/v3/reference/teamdrives/create}
}
}
