% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_drive_get.R
\name{team_drive_get}
\alias{team_drive_get}
\title{Get Team Drives by name or id}
\usage{
team_drive_get(name = NULL, id = NULL, verbose = TRUE)
}
\arguments{
\item{name}{Character vector of names.  A character vector marked with
\code{\link[=as_id]{as_id()}} is treated as if it was provided via the \code{id} argument.}

\item{id}{Character vector of Team Drive ids or URLs (it is first processed
with \code{\link[=as_id]{as_id()}}). If both \code{name} and \code{id} are non-\code{NULL}, \code{id} is silently
ignored.}

\item{verbose}{Logical, indicating whether to print informative messages
(default \code{TRUE}).}
}
\value{
An object of class \code{\link{dribble}}, a tibble with one row per item.
}
\description{
Retrieve metadata for Team Drives specified via name or id. Note
that Google Drive does NOT behave like your local file system:
\itemize{
\item You can get zero, one, or more Team Drives back for each name! Team Drive
names need not be unique.
}

Team Drives have been rebranded as \emph{shared drives} and, as of googledrive
v2.0.0, all \verb{team_drive_*()} functions have been deprecated, in favor of
\verb{shared_drive_*()} successors.

The changes in googledrive reflect that the Team Drives resource collection
has been deprecated in the Drive API v3, in favor of the new (shared) Drives
resource collection. Read more
\itemize{
\item \url{https://cloud.google.com/blog/products/application-development/upcoming-changes-to-the-google-drive-api-and-google-picker-api}
}
}
\examples{
\dontrun{
team_drive_get("my-awesome-team-drive")
team_drive_get(c("apple", "orange", "banana"))
team_drive_get(as_id("KCmiHLXUk9PVA-0AJNG"))
team_drive_get(as_id("https://drive.google.com/drive/u/0/folders/KCmiHLXUk9PVA-0AJNG"))
team_drive_get(id = "KCmiHLXUk9PVA-0AJNG")
team_drive_get(id = "https://drive.google.com/drive/u/0/folders/KCmiHLXUk9PVA-0AJNG")
}
}
