% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_corpus.R
\name{teamdrives}
\alias{teamdrives}
\title{Access a Team Drive}
\description{
The Drive API requires extra information in order to look for
files on a Team Drive. \code{\link[=drive_find]{drive_find()}} and \code{\link[=drive_get]{drive_get()}} accept this
information via \code{...} and should be used to capture Team Drives files in a
\code{\link{dribble}}, suitable for use in other googledrive functions.

Note:
  \href{https://gsuite.google.com/learning-center/products/drive/get-started-team-drive/}{Team Drives}
  are only available to users of certain enhanced Google services, such as G
  Suite Enterprise, G Suite Business, or G Suite for Education.

If you want to search a specific Team Drive, provide its id
somewhere in the call, like so:\preformatted{drive_find(..., teamDriveId = "XXXXXXXXX")
}

To search other collections, pass the \code{corpora} parameter somewhere in the
call, like so:\preformatted{drive_find(..., corpora = "user,allTeamDrives")
}

Possible values of \code{corpora} and what they mean:
\itemize{
\item \code{"user"}: Queries files that the user has accessed, including both Team
and non-Team Drive files.
\item \code{"teamDrive"}: Queries all items in one specific Team Drive.
\code{teamDriveId} must be also specified in the request. In fact, googledrive
infers the need to send \code{corpora = "teamDrive"} whenever \code{teamDriveId} is
set.
\item \code{"user,allTeamDrives"}: Queries files that the user has accessed and all
Team Drives in which they are a member. Note that both \code{"user"} and
\code{"teamDrive"} are preferable to \code{"allTeamDrives"}, in terms of efficiency.
\item \code{"domain"}: Queries files that are shared to the domain, including both
Team Drive and non-Team Drive files.
}
}
\details{
Team Drive support:
googledrive implements Team Drive support as outlined here:
\itemize{
\item \url{https://developers.google.com/drive/v3/web/enable-teamdrives#including_team_drive_content_fileslist}
}

Users shouldn't need to know any of this, but here are details for the
curious. The extra information needed to search Team Drives consists of:
\itemize{
\item \code{corpora}: Where to search? Described above.
\item \code{teamDriveId}: The id of a specific Team Drive. Only allowed -- and also
absolutely required -- when \code{corpora = "teamDrive"}. When user passes a
\code{teamDriveId}, googledrive sends it and also infers that \code{corpora} should
be set to \code{"teamDrive"} and sent.
\item \code{includeTeamDriveItems}: Do you want to see Team Drive items? Obviously,
this should be \code{TRUE} and googledrive sends this whenever Team Drive
parameters are detected
\item \code{supportsTeamDrives}: Does the sending application (googledrive, in this
case) know about Team Drives? Obviously, this should be \code{TRUE} and
googledrive sends it for all applicable endpoints, all the time.
}
}
\seealso{
\code{\link[=drive_find]{drive_find()}}
}
