% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_id.R
\name{as_id}
\alias{as_id}
\title{Mark as Google Drive id.}
\usage{
as_id(x)
}
\arguments{
\item{x}{Character vector of Google Drive ids or Google Drive URLs}
}
\description{
Marks a character vector as holding Google Drive file ids, as opposed to file
names or paths.
}
