% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{copy_file}
\alias{copy_file}
\title{Get file via name}
\usage{
copy_file(fileID, folderID = NULL, fileName = NULL)
}
\arguments{
\item{fileID}{ID of the file in Google Drive}

\item{folderID}{ID of the folder to store the copied file}

\item{fileName}{The name of the file. This is not necessarily unique within a folder.}
}
\description{
Get file via name
}

