% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{copy_file}
\alias{copy_file}
\title{Copy a file in Google Drive}
\usage{
copy_file(file_id, folder_id = "root", file_name = NULL)
}
\arguments{
\item{file_id}{ID of the file in Google Drive}

\item{folder_id}{ID of the folder to store the copied file}

\item{file_name}{The name to be given to the copied file. This does not need to be unique within a folder.}
}
\description{
Copy a file in Google Drive
}

