% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dribble.R
\name{dribble}
\alias{dribble}
\title{dribble object}
\description{
googledrive stores the metadata for one or more Google Drive files as a
\code{dribble}. It is a \link[tibble:tibble-package]{tibble} with one row per file
and, at a minimum, character variables containing file name and id and a
list-column of \href{https://developers.google.com/drive/v3/reference/files#resource}{Files resource}
objects (possibly incomplete).
}
