% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_request.R
\name{drive_api_key}
\alias{drive_api_key}
\title{Get an API key}
\usage{
drive_api_key(key = NULL)
}
\arguments{
\item{key}{Character, optional. A Google API key.}
}
\value{
A Google API key
}
\description{
Pass through an API key that is explicitly provided. Otherwise, consult the
environment variable \code{GOOGLEDRIVE_API_KEY} and the API key built-into
googledrive, in that order.
}
\examples{
## specify explicitly
drive_api_key("I_have_my_own_key")

## specify via env var
tryCatch({
  Sys.setenv(GOOGLEDRIVE_API_KEY = "a1b2c3d4e5f7")
  drive_api_key()
  },
  finally = Sys.unsetenv("GOOGLEDRIVE_API_KEY")
)

## use the built-in API key
drive_api_key()
}
