% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_auth.R
\name{drive_auth_config}
\alias{drive_auth_config}
\title{View or set auth config}
\usage{
drive_auth_config(active = TRUE, app = NULL, api_key = NULL,
  verbose = TRUE)
}
\arguments{
\item{active}{Logical. \code{TRUE} means a token will be sent. \code{FALSE} means it
will not.}

\item{app}{OAuth app. Defaults to tidyverse app that ships with googledrive.}

\item{api_key}{API key. Defaults to key that ships with googledrive.
Necessary in order to make unauthorized "token-free" requests for public
resources.}

\item{verbose}{Logical, indicating whether to print informative messages
(default \code{TRUE}).}
}
\value{
\code{NULL}, invisibly
}
\description{
If you want to use your own app, setup a new project in
\href{https://console.developers.google.com}{Google Developers Console}.
}
\examples{
drive_auth_config()
}
