% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_auth.R
\name{drive_auth_config}
\alias{drive_auth_config}
\alias{drive_api_key}
\alias{drive_oauth_app}
\title{View or edit auth config}
\usage{
drive_auth_config(app = NULL, path = NULL, api_key = NULL)

drive_api_key()

drive_oauth_app()
}
\arguments{
\item{app}{OAuth app. Defaults to a tidyverse app.}

\item{path}{Path to the JSON file.}

\item{api_key}{API key. Defaults to a tidyverse key. Necessary in
order to make unauthorized "token-free" requests for public resources.}
}
\value{
\code{drive_auth_config()}: An object of R6 class \code{AuthState},
which is defined in the gargle package. \code{drive_api_key()}: the
current API key. \code{drive_oauth_app()}: the current
\code{\link[httr:oauth_app]{httr::oauth_app()}}.
}
\description{
These functions give the user more control over auth than what is
possible with \code{\link[=drive_auth]{drive_auth()}}. \code{drive_auth_config()} gives control
of:
\itemize{
\item The OAuth app, which is used when obtaining a user token.
\item The API key. If googledrive is deauthorized via \code{\link[=drive_deauth]{drive_deauth()}},
all requests will be sent with an API key in lieu of a token.
}

See the vignette \href{https://gargle.r-lib.org/articles/get-api-credentials.html}{How to get your own API credentials}
for more.

\code{drive_api_key()} and \code{drive_oauth_app()} retrieve the
currently configured API key and OAuth app, respectively.
}
\examples{
## this will print current config
drive_auth_config()

if (require(httr)) {
  ## bring your own app via client id (aka key) and secret
  google_app <- httr::oauth_app(
    "my-awesome-google-api-wrapping-package",
    key = "123456789.apps.googleusercontent.com",
    secret = "abcdefghijklmnopqrstuvwxyz"
  )
  google_key <- "the-key-I-got-for-a-google-API"
  drive_auth_config(app = google_app, api_key = google_key)
}

\dontrun{
## bring your own app via JSON downloaded from Google Developers Console
drive_auth_config(
  path = "/path/to/the/JSON/you/downloaded/from/google/dev/console.json"
)
}
}
\seealso{
Other auth functions: \code{\link{drive_auth}},
  \code{\link{drive_deauth}}
}
\concept{auth functions}
