% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_browse.R
\name{drive_browse}
\alias{drive_browse}
\title{Visit Google Drive file in browser.}
\usage{
drive_browse(file)
}
\arguments{
\item{file}{Drive File, something that identifies the file of interest
on your Google Drive. This can be name(s)/path(s), marked file id(s),
or a \code{dribble}.}
}
\value{
An object of class \code{dribble}, a tibble with
   one row per file.
}
\description{
Visit Google Drive file in browser.
}
