% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_browse.R
\name{drive_browse}
\alias{drive_browse}
\title{Open Google Drive file in browser to edit}
\usage{
drive_browse(file)
}
\arguments{
\item{file}{Drive File, something that identifies the file of interest
on your Google Drive. This can be name(s)/path(s), marked file id(s),
or a `dribble`.}
}
\value{
Drive File. An object of class `dribble`, a tibble with
   one row per file.
}
\description{
Open Google Drive file in browser to edit
}
