% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_browse.R
\name{drive_browse}
\alias{drive_browse}
\title{Visit Drive file in browser.}
\usage{
drive_browse(file = .Last.value)
}
\arguments{
\item{file}{Something that identifies the file of interest on your Google
Drive. Can be a character vector of names/paths, a character vector of file
ids or URLs marked with \code{\link{as_id}()}, or a \code{\link{dribble}}.}
}
\value{
An object of class \code{\link{dribble}}, a tibble with one row per
  item.
}
\description{
Visit Drive file in browser.
}
