% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_browse.R
\name{drive_browse}
\alias{drive_browse}
\title{Visit Drive file in browser}
\usage{
drive_browse(file = .Last.value)
}
\arguments{
\item{file}{Something that identifies the file of interest on your Google
Drive. Can be a name or path, a file id or URL marked with \code{\link[=as_id]{as_id()}}, or a
\code{\link{dribble}}.}
}
\value{
Character vector of file hyperlinks, from \code{\link[=drive_link]{drive_link()}}, invisibly.
}
\description{
Visits a file on Google Drive in your default browser.
}
\examples{
\dontshow{if (drive_has_token() && rlang::is_interactive()) withAutoprint(\{ # examplesIf}
drive_find(n_max = 1) \%>\% drive_browse()
\dontshow{\}) # examplesIf}
}
