% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_corpus.R
\name{drive_corpus}
\alias{drive_corpus}
\title{Search for Team Drive content}
\usage{
drive_corpus(corpora = NULL, teamDriveId = NULL,
  includeTeamDriveItems = NULL)
}
\arguments{
\item{corpora}{Character. One of \code{"user"}, \code{"teamDrive"},
\code{"user,allTeamDrives"}, or \code{"domain"}}

\item{teamDriveId}{Character. Supply a Team Drive id here if and only if
\code{corpora = "teamDrive"}.}

\item{includeTeamDriveItems}{Logical. Should always be \code{TRUE} when addressing
Team Drives and, indeed, googledrive will make it so before calling the
API.}
}
\value{
A list with class \code{drive_corpus}.
}
\description{
The Drive API requires extra information if you want to search
for files within a specific Team Drive or across all Team Drives. This
matters for more functions than you might think, since many googledrive
calls do some file searching behind the scenes, for example, whenever there
is a need to resolve a file id based on a file's name or path.

The extra information consists of:
\itemize{
\item \code{corpora}: Where to search?
\item \code{teamDriveId}: The id of a specific Team Drive. Only relevant if
\code{corpora = "teamDrive"}.
}

Possible values of \code{corpora} and what they mean:
\itemize{
\item \code{"user"}: Queries files that the user has accessed, including both Team
and non-Team Drive files.
\item \code{"teamDrive"}: Queries all items in one specific Team Drive.
\code{teamDriveId} must be also specified in the request.
\item \code{"user,allTeamDrives"}: Queries files that the user has accessed and all
Team Drives in which they are a member. Prefer \code{"user"} or \code{"teamDrive"}
to \code{"allTeamDrives"} for efficiency.
\item \code{"domain"}: Queries files that are shared to the domain, including both
Team Drive and non-Team Drive files.
}

googledrive adds \code{includeTeamDriveItems = TRUE} to the query
whenever \code{corpora} is specified. The user does not need to specify this
and indeed should not.
}
\examples{
drive_corpus("user")
}
\seealso{
Implements Team Drive access as described here:
\itemize{
\item \url{https://developers.google.com/drive/v3/web/enable-teamdrives#including_team_drive_content_fileslist}
}
}
