% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_cp.R
\name{drive_cp}
\alias{drive_cp}
\alias{drive_copy}
\title{Copy a Google Drive file.}
\usage{
drive_cp(file = NULL, verbose = TRUE)

drive_copy(file = NULL, verbose = TRUE)
}
\arguments{
\item{file}{Drive File, something that identifies the file of interest
on your Google Drive. This can be name(s)/path(s), marked file id(s),
or a \code{dribble}.}

\item{verbose}{Logical, indicating whether to print informative messages
(default \code{TRUE}).}
}
\value{
An object of class \code{\link{dribble}}, a tibble with
   one row per file.
}
\description{
\code{drive_copy()} is an alias for \code{drive_cp()}.
}
\examples{
\dontrun{
## Copy the file "chickwts.rda" in the same folder as the
## original file as "Copy of chickwts.rda".
drive_cp("chickwts.rda")

## Copy multiple files - this will create files
## "Copy of chickwts.rda" and "Copy of chickwts.csv"
## in their same folder as the respective original files.
drive_cp(c("chickwts.rda", "chickwts.csv"))
}
}
