% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_cp.R
\name{drive_cp}
\alias{drive_cp}
\title{Copy a Drive file.}
\usage{
drive_cp(file, path = NULL, name = NULL, verbose = TRUE)
}
\arguments{
\item{file}{Something that identifies the file(s) of interest on your Google
Drive. Can be a character vector of names/paths, a character vector of file
ids marked with \code{\link{as_id}()}, or a \code{\link{dribble}}.}

\item{path}{Specifies target destination for the new file on Google Drive.
Can be an actual path (character), a file id marked with
\code{\link{as_id}()}, or a \code{\link{dribble}}. If specified as an
actual path, it is best to explicitly indicate if it's a folder by
including a trailing slash, since it cannot always be worked out from the
context of the call.}

\item{name}{Character, new file  name if not specified as part of `path`. This
will force `path` to be treated as a folder, even if it is character and lacks
a trailing slash. Defaults to "Copy of \code{FILE-NAME}".}

\item{verbose}{Logical, indicating whether to print informative messages
(default \code{TRUE}).}
}
\value{
An object of class \code{\link{dribble}}, a tibble with one row per
  file.
}
\description{
Copy a Drive file.
}
\examples{
\dontrun{
## create a file to copy
file <- drive_upload(system.file("DESCRIPTION"), "DESC-ex")

## Make a "Copy of" copy in same folder as the original
drive_cp("DESC-ex")

## Make an explicitly named copy in same folder as the original
drive_cp("DESC-ex", "DESC-ex-two")

## Make an explicitly named copy in a different folder
folder <- drive_mkdir("new-folder")
drive_cp("DESC-ex", folder, "DESC-ex-three")

## Behold all of our copies!
drive_find("DESC-ex")

## Delete all of our copies and the new folder!
drive_find("DESC-ex") \%>\% drive_rm()
drive_rm(folder)
}
}
\seealso{
Wraps the
\href{https://developers.google.com/drive/v3/reference/files/copy}{drive.files.copy}
endpoint
}
