% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_create.R
\name{drive_create}
\alias{drive_create}
\title{Create a new blank Drive file}
\usage{
drive_create(name, parent = NULL, type = NULL, ..., verbose = TRUE)
}
\arguments{
\item{name}{Name for the new file or, optionally, a path that specifies
an existing parent folder, as well as the new file name.}

\item{parent}{Target destination for the new file, i.e. a folder or a Team
Drive. Can be given as an actual path (character), a file id, or URL marked
with \code{\link[=as_id]{as_id()}}, or a \code{\link{dribble}}. Defaults to your "My Drive" root folder.}

\item{type}{Character. Create a blank Google Doc, Sheet or Slides by
setting \code{type} to \code{document}, \code{spreadsheet}, or \code{presentation},
respectively. All non-\code{NULL} values for \code{type} are pre-processed with
\code{\link[=drive_mime_type]{drive_mime_type()}}.}

\item{...}{Named parameters to pass along to the Drive API. You can affect
the metadata of the target file by specifying properties of the Files
resource via \code{...}. See the "Request body" section of the Drive API
docs for the associated endpoint.}

\item{verbose}{Logical, indicating whether to print informative messages
(default \code{TRUE}).}
}
\value{
An object of class \code{\link{dribble}}, a tibble with one row per item.
}
\description{
Creates a new blank Drive file. To create a new Drive \emph{folder},
use \code{\link[=drive_mkdir]{drive_mkdir()}}. To upload an existing local file into
Drive, use \code{\link[=drive_upload]{drive_upload()}}.
}
\examples{
\dontrun{
## Create a blank Google Doc named 'Word Star' in
## your 'My Drive' root folder and star it
wordstar <- drive_create("WordStar", type = "document", starred = TRUE)

## is 'WordStar' really starred? YES
purrr::pluck(wordstar, "drive_resource", 1, "starred")

## Create a blank Google Slides presentation in
## the root folder, and set its description
execuvision <- drive_create(
  "ExecuVision",
  type = "presentation",
  description = "deeply nested bullet lists FTW"
)

## Did we really set the description? YES
purrr::pluck(execuvision, "drive_resource", 1, "description")

## check out the new presentation
drive_browse(execuvision)

## Create folder 'b4xl' in the root folder,
## then create an empty new Google Sheet in it
b4xl <- drive_mkdir("b4xl")
drive_create("VisiCalc", parent = b4xl, type = "spreadsheet")

## Another way to create a Google Sheet in the folder 'spreadsheets'
drive_create("b4xl/SuperCalc", type = "spreadsheet")

## Another way to create a new file in a folder,
## this time specifying `parent` as a character
drive_create("Lotus 1-2-3", parent = "b4xl", type = "spreadsheet")

## clean up
drive_rm(wordstar, b4xl, execuvision)
}
}
\seealso{
Wraps the \code{files.create} endpoint:
\itemize{
\item \url{https://developers.google.com/drive/v3/reference/files/create}
}
}
