% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_create.R
\name{drive_create}
\alias{drive_create}
\title{Create a new blank Drive file}
\usage{
drive_create(
  name,
  path = NULL,
  type = NULL,
  ...,
  overwrite = NA,
  verbose = TRUE
)
}
\arguments{
\item{name}{Name for the new file or, optionally, a path that specifies
an existing parent folder, as well as the new file name.}

\item{path}{Target destination for the new item, i.e. a folder or a Team
Drive. Can be given as an actual path (character), a file id or URL marked
with \code{\link[=as_id]{as_id()}}, or a \code{\link{dribble}}. Defaults to your "My Drive" root folder.}

\item{type}{Character. Create a blank Google Doc, Sheet or Slides by
setting \code{type} to \code{document}, \code{spreadsheet}, or \code{presentation},
respectively. All non-\code{NULL} values for \code{type} are pre-processed with
\code{\link[=drive_mime_type]{drive_mime_type()}}.}

\item{...}{Named parameters to pass along to the Drive API. Has the tidy dots
semantics that come from using \code{\link[rlang:list2]{rlang::list2()}}. You can affect the
metadata of the target file by specifying properties of the Files resource
via \code{...}. Read the "Request body" section of the Drive API docs for the
associated endpoint to learn about relevant parameters.}

\item{overwrite}{Logical, indicating whether to check for a pre-existing file
at the targetted "filepath". The quotes around "filepath" refer to the fact
that Drive does not impose a 1-to-1 relationship between filepaths and files,
like a typical file system; read more about that in \code{\link[=drive_get]{drive_get()}}.
\itemize{
\item \code{NA} (default): Just do the operation, even if it results in multiple
files with the same filepath.
\item \code{TRUE}: Check for a pre-existing file at the filepath. If there is
zero or one, move a pre-existing file to the trash, then carry on. Note
that the new file does not inherit any properties from the old one, such
as sharing or publishing settings. It will have a new file ID. An error is
thrown if two or more pre-existing files are found.
\item \code{FALSE}: Error if there is any pre-existing file at the filepath.
}

Note that existence checks, based on filepath, are expensive operations, i.e.
they require additional API calls.}

\item{verbose}{Logical, indicating whether to print informative messages
(default \code{TRUE}).}
}
\value{
An object of class \code{\link{dribble}}, a tibble with one row per item.
}
\description{
Creates a new blank Drive file. Note there are better options for these
special cases:
\itemize{
\item Creating a folder? Use \code{\link[=drive_mkdir]{drive_mkdir()}}.
\item Want to upload existing local content into a new Drive file? Use
\code{\link[=drive_upload]{drive_upload()}}.
}
}
\examples{
\dontrun{
## Create a blank Google Doc named 'WordStar' in
## your 'My Drive' root folder and star it
wordstar <- drive_create("WordStar", type = "document", starred = TRUE)

## is 'WordStar' really starred? YES
purrr::pluck(wordstar, "drive_resource", 1, "starred")

## Create a blank Google Slides presentation in
## the root folder, and set its description
execuvision <- drive_create(
  "ExecuVision",
  type = "presentation",
  description = "deeply nested bullet lists FTW"
)

## Did we really set the description? YES
purrr::pluck(execuvision, "drive_resource", 1, "description")

## check out the new presentation
drive_browse(execuvision)

## Create folder 'b4xl' in the root folder,
## then create an empty new Google Sheet in it
b4xl <- drive_mkdir("b4xl")
drive_create("VisiCalc", path = b4xl, type = "spreadsheet")

## Another way to create a Google Sheet in the folder 'b4xl'
drive_create("b4xl/SuperCalc", type = "spreadsheet")

## Another way to create a new file in a folder,
## this time specifying parent `path` as a character
drive_create("Lotus 1-2-3", path = "b4xl", type = "spreadsheet")

## `overwrite = FALSE` errors if file already exists at target filepath
## THIS WILL ERROR!
drive_create("VisiCalc", path = b4xl, overwrite = FALSE)

## `overwrite = TRUE` moves an existing file to trash, then proceeds
drive_create("VisiCalc", path = b4xl, overwrite = TRUE)

## clean up
drive_rm(wordstar, b4xl, execuvision)
}
}
\seealso{
Wraps the \code{files.create} endpoint:
\itemize{
\item \url{https://developers.google.com/drive/v3/reference/files/create}
}
}
