% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_auth.R
\name{drive_deauth}
\alias{drive_deauth}
\title{Suspend authorization}
\usage{
drive_deauth(clear_cache = TRUE, verbose = TRUE)
}
\arguments{
\item{clear_cache}{logical indicating whether to disable the
\code{.httr-oauth} file in working directory, if such exists, by renaming
to \code{.httr-oauth-SUSPENDED}}

\item{verbose}{Logical, indicating whether to print informative messages
(default \code{TRUE}).}
}
\description{
Suspend googledrive's authorization to place requests to the Drive API on
behalf of the authenticated user.
}
\examples{
\dontrun{
drive_deauth()
}
}
\seealso{
Other auth functions: \code{\link{drive_auth_config}},
  \code{\link{drive_auth}}
}
