% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_delete.R
\name{drive_delete}
\alias{drive_delete}
\title{Delete file from Drive.}
\usage{
drive_delete(file = NULL, verbose = TRUE)
}
\arguments{
\item{file}{Drive File, something that identifies the file of interest
on your Google Drive. This can be name(s)/path(s), marked file id(s),
or a \code{dribble}.}

\item{verbose}{Logical, indicating whether to print informative messages
(default \code{TRUE}).}
}
\value{
Logical, indicating whether the delete succeeded.
}
\description{
Delete file from Drive.
}
