% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_download.R
\name{drive_download}
\alias{drive_download}
\title{Download a file from Google Drive.}
\usage{
drive_download(file = NULL, out_path = NULL, overwrite = FALSE,
  verbose = TRUE)
}
\arguments{
\item{file}{Drive File, something that identifies the file of interest
on your Google Drive. This can be name(s)/path(s), marked file id(s),
or a \code{dribble}.}

\item{out_path}{Character. Path for output file. Will default to its Google Drive
name. Google files must include the intended extension and can be downloaded to types
specified in the \href{https://developers.google.com/drive/v3/web/manage-downloads}{Drive API documentation}.}

\item{overwrite}{A logical scalar, do you want to overwrite a file that already
exists locally?}

\item{verbose}{Logical, indicating whether to print informative messages
(default \code{TRUE}).}
}
\description{
Download a file from Google Drive.
}
\examples{
\dontrun{
## Save "chickwts.csv" to the working directory as "chickwts.csv".
drive_download(file = "chickwts.csv")

## Export a Google Sheet named "chickwts" to the working directory as
## "chickwts.csv".
drive_download(file = "chickwts", out_path = "chickwts.csv")

## Export a Google Document named "foobar" to the working directory as
## "foobar.docx"
drive_download(file = "foobar", out_path = "foobar.docx")
}
}
