% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_trash.R
\name{drive_empty_trash}
\alias{drive_empty_trash}
\title{Empty Drive Trash.
Caution: this will permanently delete files in your Drive trash.}
\usage{
drive_empty_trash(verbose = TRUE)
}
\arguments{
\item{verbose}{Logical, indicating whether to print informative messages
(default \code{TRUE}).}
}
\description{
Empty Drive Trash.
Caution: this will permanently delete files in your Drive trash.
}
