% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_fields.R
\name{drive_fields}
\alias{drive_fields}
\alias{prep_fields}
\title{Request partial resources}
\usage{
drive_fields(fields = NULL, resource = "files", which = c("default",
  "all"))

prep_fields(fields, resource = "files")
}
\arguments{
\item{fields}{Character vector of field names. Optional.}

\item{resource}{Character, naming the API resource of interest.}

\item{which}{Consulted only when \code{resource = "files"} and \code{fields = NULL},
dictates whether all fields are returned or the googledrive defaults.}
}
\value{
Character vector of field names or a single string.
}
\description{
You may be able to improve the performance of your API calls by requesting
only the metadata that you actually need. This function is primarily for
internal use and is geared towards the
\href{https://developers.google.com/drive/v3/reference/files}{Files resource}.
Assuming that \code{resource = "files"}, \code{drive_fields()} returns the \code{default}
fields, \code{all} fields, or, if \code{fields} is non-\code{NULL}, the subset of the input
fields that are recognized as valid fields. \code{prep_fields()} prepares fields
for inclusion as query parameters.
}
\examples{
## to get default fields or all fields for the Files resource
drive_fields()
drive_fields(which = "all")

## invalid fields are removed and throw warning
drive_fields(c("name", "parents", "ownedByMe", "pancakes!"))

## fields for query
prep_fields(c("name", "parents", "ownedByMe"))
}
\seealso{
\href{https://developers.google.com/drive/v3/web/performance#partial}{Working with partial resources},
in the Drive API documentation
}
