% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_get.R
\name{drive_get}
\alias{drive_get}
\title{Get Drive files by id}
\usage{
drive_get(id)
}
\arguments{
\item{id}{Character vector of Drive file ids, such as you might see in the
URL when visiting a file on Google Drive.}
}
\value{
dribble-return
}
\description{
Warning: the name of this function is likely to change.
}
\examples{
\dontrun{
drive_get("abcdefgeh123456789")

drive_get(c("abcdefgh123456789", "jklmnopq123456789"))

## the id "root" is an alias for your root folder on My Drive
drive_get("root")
}
}
