% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_get.R
\name{drive_get}
\alias{drive_get}
\title{Get Drive files by path or id}
\usage{
drive_get(path = NULL, id = NULL, team_drive = NULL, corpora = NULL,
  verbose = TRUE)
}
\arguments{
\item{path}{Character vector of path(s) to get. Use a trailing slash to
indicate explicitly that a path is a folder, which can disambiguate if
there is a file of the same name (yes this is possible on Drive!). A
character vector marked with \code{\link[=as_id]{as_id()}} is treated as if it was provided via
the \code{id} argument.}

\item{id}{Character vector of Drive file ids or URLs (it is first processed
with \code{\link[=as_id]{as_id()}}). If both \code{path} and \code{id} are non-\code{NULL}, \code{id} is silently
ignored.}

\item{team_drive}{Anything that identifies one specific Team Drive: its name,
its id or URL marked with \code{\link{as_id}()}, or a
\code{\link{dribble}}. Is pre-processed with
\code{\link{as_teamdrive}()} and then \code{\link{drive_corpus}()}.}

\item{corpora}{Character, specifying the search collection. Only relevant in
the Team Drives context.  Is pre-processed together with \code{team_drive}
by \code{\link{drive_corpus}()}.}

\item{verbose}{Logical, indicating whether to print informative messages
(default \code{TRUE}).}
}
\value{
An object of class \code{\link{dribble}}, a tibble with one row per
  item.
}
\description{
Retrieve metadata for files specified via path or via file id.
Note that Google Drive does NOT behave like your local file system:
\itemize{
\item You can get zero, one, or more files back for each file name or path! On
Google Drive, file and folder names need not be unique, even at a given
level of the hierarchy.
\item A file or folder can also have multiple direct parents, so a single Drive
file can potentially be represented by multiple paths.
}

In contrast, a file id will always identify at most one Drive
file. Finally, note also that a folder is just a specific type of file on
Drive.

If the files are specified via \code{path}, versus \code{id}, the returned \code{\link{dribble}}
will include a \code{path} variable. To add path information to any \code{\link{dribble}}
that lacks it, use \code{\link[=drive_show_path]{drive_show_path()}}. If you want to list the contents of a
folder, use \code{\link[=drive_ls]{drive_ls()}}. For general searching, use \code{\link[=drive_find]{drive_find()}}.

Note:
  \href{https://gsuite.google.com/learning-center/products/drive/get-started-team-drive/}{Team Drives}
  are only available to users of certain enhanced Google services, such as G
  Suite Enterprise, G Suite Business, or G Suite for Education.
}
\examples{
\dontrun{
## get info about your "My Drive" root folder
drive_get("~/")
## the API reserves the file id "root" for your root folder
drive_get(id = "root")
drive_get(id = "root") \%>\% drive_show_path()

## multiple names
drive_get(c("abc", "def"))

## multiple names, one of which must be a folder
drive_get(c("abc", "def/"))

## query by file id(s)
drive_get(id = "abcdefgeh123456789")
drive_get(as_id("abcdefgeh123456789"))
drive_get(id = c("abcdefgh123456789", "jklmnopq123456789"))

## access the Team Drive named "foo"
foo <- teamdrive_get("foo")
drive_get(c("this.jpg", "that-file"), team_drive = foo)
drive_get(as_id("123456789"), team_drive = foo)

## search all Team Drives and other files user has accessed
drive_get(c("this.jpg", "that-file"), corpora = "user,allTeamDrives")
}
}
\seealso{
Wraps the \code{files.get} endpoint and, if you specify files by name or
path, also calls \code{files.list}:
\itemize{
\item \url{https://developers.google.com/drive/v3/reference/files/update}
\item \url{https://developers.google.com/drive/v3/reference/files/list}
}
}
