% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_auth.R
\name{drive_has_token}
\alias{drive_has_token}
\title{Is there a token on hand?}
\usage{
drive_has_token()
}
\value{
Logical.
}
\description{
Reports whether googledrive has stored a token, ready for use in downstream
requests. Exists mostly for protecting examples that won't work in the
absence of a token.
}
\examples{
drive_has_token()
}
