% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_id.R
\name{drive_id}
\alias{drive_id}
\title{Register Drive id}
\usage{
drive_id(x)
}
\arguments{
\item{x}{character, drive id to register}
}
\description{
This addes the class \code{drive_id} to a character string.
}
