% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_publish.R
\name{drive_is_published}
\alias{drive_is_published}
\title{Check if Google Drive file is published}
\usage{
drive_is_published(file, verbose = TRUE)
}
\arguments{
\item{file}{Something that identifies the file(s) of interest on your Google
Drive. Can be a character vector of names/paths, a character vector of file
ids marked with \code{\link{as_id}()}, or a \code{\link{dribble}}.}

\item{verbose}{Logical, indicating whether to print informative messages
(default \code{TRUE}).}
}
\value{
An object of class \code{\link{dribble}}, a tibble with one row per
  file.
}
\description{
Check if Google Drive file is published
}
\examples{
## Upload file to check publication status
file <- drive_upload(R.home('doc/BioC_mirrors.csv'),
  type = "spreadsheet")

## Check publication status
drive_is_published(file)

## Publish file
drive_publish(file)

## Check publication status agian
drive_is_published(file)

## Unpublish file
drive_unpublish(file)

## Clean up
drive_rm(file)
}
