% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_list.R
\name{drive_list}
\alias{drive_list}
\title{List files on Google Drive}
\usage{
drive_list(path = NULL, pattern = NULL, ..., verbose = TRUE)
}
\arguments{
\item{path}{character, path where the Google drive files are
that you would like to list. Defaults to the "My Drive" directory.}

\item{pattern}{character, regular expression of titles of
documents to output in a tibble. If it is \code{NULL} (default), information
about all documents in drive will be output in a tibble.}

\item{...}{name-value pairs to query the API}

\item{verbose}{logical, indicating whether to print informative messages
(default \code{TRUE})

This will default to the most recent 100 files on your Google Drive. If you
would like more than 100, include the \code{pageSize} parameter. For example, if
I wanted 200, I would run \code{drive_ls(pageSize = 200)}.
Helpful links for forming queries:
\itemize{
\item \url{https://developers.google.com/drive/v3/web/search-parameters}
\item \url{https://developers.google.com/drive/v3/reference/files/list}
}}
}
\value{
tibble containing the name, type, and id of files on your google
drive (default 100 files)
}
\description{
List files on Google Drive
}
\examples{
\dontrun{
## list user's Google Sheets
drive_list(q = "mimeType='application/vnd.google-apps.spreadsheet'")
}

}
