% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_list.R
\name{drive_list}
\alias{drive_list}
\title{List files on Google Drive}
\usage{
drive_list(path = NULL, pattern = NULL, ..., verbose = TRUE)
}
\arguments{
\item{path}{character. Google Drive path to list. Defaults to "My Drive", but
without regard to any folder hierarchy. If path is a folder, its contents
are listed, not recursively. If path is a file, that single file is listed.
A trailing slash indicates explicitly that the path is a folder, which can
disambiguate if there is a file of the same name (yes this is possible on
Drive!).}

\item{pattern}{character. If provided, only the files whose names match this
regular expression are returned.}

\item{...}{Parameters to pass along to the API query.}

\item{verbose}{logical. Indicates whether to print informative messages.

This will default to the most recent 100 files on your Google Drive. For
example, to get 200 instead, specify the \code{pageSize}, i.e.
\code{drive_ls(pageSize = 200)}.
Helpful links for forming queries:
\itemize{
\item \url{https://developers.google.com/drive/v3/web/search-parameters}
\item \url{https://developers.google.com/drive/v3/reference/files/list}
}}
}
\value{
tibble with one row per file
}
\description{
List files on Google Drive
}
\examples{
\dontrun{
## list "My Drive" w/o regard for folders
drive_list()

## list files that specifically live at the top-level of "My Drive"
drive_list("~/")

## just folders
drive_list(q = "mimeType = 'application/vnd.google-apps.folder'")

## just folders that have the folder with id 'abc' as direct parent
drive_list(q = "'abc' in parents and mimeType='application/vnd.google-apps.folder'")

## files that match a regex
drive_list(pattern = "jt")

## list the contents of the 'jt01' folder
drive_list("jt01/")

## list user's Google Sheets
drive_list(q = "mimeType='application/vnd.google-apps.spreadsheet'")
}

}
