% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_ls.R
\name{drive_ls}
\alias{drive_ls}
\title{List files on Google Drive}
\usage{
drive_ls(search = NULL, ..., fixed = FALSE, verbose = TRUE)
}
\arguments{
\item{search}{character, regular expression(s) of title(s) of documents to
output in a tibble. If it is \code{NULL} (default), information about all
documents in drive will be output in a tibble.}

\item{...}{name-value pairs to query the API}

\item{fixed}{logical, from \code{\link[=grep]{grep()}}. If \code{TRUE}, \code{search} is exactly matched
to a document's name on Google Drive.}

\item{verbose}{logical, indicating whether to print informative messages
(default \code{TRUE})

This will default to the most recent 100 files on your Google Drive. If you
would like more than 100, include the \code{pageSize} parameter. For example, if I
wanted 200, I would run \code{drive_ls(pageSize = 200)}.

Helpful links for forming queries:
\itemize{
\item \url{https://developers.google.com/drive/v3/web/search-parameters}
\item \url{https://developers.google.com/drive/v3/reference/files/list}
}}
}
\value{
tibble containing the name, type, and id of files on your google
drive (default 100 files)
}
\description{
List files on Google Drive
}
\examples{
\dontrun{
## list user's Google Sheets
drive_ls(q = "mimeType='application/vnd.google-apps.spreadsheet'")
}

}
