% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_ls.R
\name{drive_ls}
\alias{drive_ls}
\title{List contents of a folder.}
\usage{
drive_ls(path = NULL, ...)
}
\arguments{
\item{path}{Specifies a single folder on Google Drive whose contents you want
to list. Can be an actual path (character), a file id or URL marked with
\code{\link[=as_id]{as_id()}}, or a \link{dribble}.}

\item{...}{Any parameters that are valid for \code{\link[=drive_find]{drive_find()}}.}
}
\value{
An object of class \code{\link{dribble}}, a tibble with one row per
  file.
}
\description{
List the contents of a folder on Google Drive, nonrecursively. This is a thin
wrapper around \code{\link[=drive_find]{drive_find()}}, that simply limits the search to a specific
folder.
}
\examples{
\dontrun{
## get contents of the folder 'abc' (non-recursive)
drive_ls("abc")

## get contents of folder 'abc' that contain the
## letters 'def'
drive_ls(path = "abc", pattern = "def")

## get all Google spreadsheets in folder 'abc'
## that contain the letters 'def'
drive_ls(path = "abc", pattern = "def", type = "spreadsheet")
}
}
