% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_mime_type.R
\name{drive_mime_type}
\alias{drive_mime_type}
\title{Lookup a mime type from a Google Drive type or file extension}
\usage{
drive_mime_type(type = NULL)
}
\arguments{
\item{type}{Character. The Google Drive type or file extension you would
like to look up the mime type of.}
}
\value{
Character. The mime type of input.
}
\description{
Lookup a mime type from a Google Drive type or file extension
}
\examples{
## get the mime type for Google Spreadsheets
drive_mime_type("spreadsheet")

## get the mime type for jpegs
drive_mime_type("jpeg")
}
