% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_mkdir.R
\name{drive_mkdir}
\alias{drive_mkdir}
\title{Create a folder on Google Drive.}
\usage{
drive_mkdir(name = NULL, folder = NULL, verbose = TRUE)
}
\arguments{
\item{name}{Character. The name of the folder you would like to create.}

\item{folder}{Drive File, something that identifies the where the file should
live on your Google Drive. This can be name(s)/path(s), marked file id(s),
or a \code{dribble}.}

\item{verbose}{Logical, indicating whether to print informative messages
(default \code{TRUE}).}
}
\value{
An object of class \code{\link{dribble}}, a tibble with
   one row per file.
}
\description{
Create a folder on Google Drive.
}
