% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_mkdir.R
\name{drive_mkdir}
\alias{drive_mkdir}
\title{Create a folder on Google Drive.}
\usage{
drive_mkdir(name = NULL, path = NULL, verbose = TRUE)
}
\arguments{
\item{name}{Character. The name of the folder you would like to create.}

\item{path}{Drive File, something that identifies the where the file should
live on your Google Drive. Can be an actual path (character), a file id marked with [as_id()], or
a [`dribble`].}

\item{verbose}{Logical, indicating whether to print informative messages
(default \code{TRUE}).}
}
\value{
An object of class \code{\link{dribble}}, a tibble with
   one row per file.
}
\description{
Create a folder on Google Drive.
}
