% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_mkdir.R
\name{drive_mkdir}
\alias{drive_mkdir}
\title{Create a folder}
\usage{
drive_mkdir(folder, parent = NULL, verbose = TRUE)
}
\arguments{
\item{folder}{Name for the new folder or, optionally, a path that specifies
an existing parent folder, as well as the new name.}

\item{parent}{Target destination for the new folder, i.e. a folder or a Team
Drive. Can be given as an actual path (character), a file id or URL marked
with \code{\link[=as_id]{as_id()}}, or a \code{\link{dribble}}. Defaults to your "My Drive" root folder.}

\item{verbose}{Logical, indicating whether to print informative messages
(default \code{TRUE}).}
}
\value{
An object of class \code{\link{dribble}}, a tibble with one row per
  item.
}
\description{
Create a folder
}
\examples{
\dontrun{
## Create folder named "def" in existing folder "abc".
drive_mkdir("abc/def")

## This will also create a folder named "def" in folder "abc".
drive_mkdir("def", parent = "abc")

## Another way to create a folder named "def" in folder "abc",
## this time with parent folder stored in a dribble.
abc <- as_dribble("abc")
drive_mkdir("def", parent = abc)

## clean up
drive_ls(path = "abc", pattern = "^def$", type = "folder") \%>\% drive_rm()
}
}
