% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_mkdir.R
\name{drive_mkdir}
\alias{drive_mkdir}
\title{Create a folder on Google Drive.}
\usage{
drive_mkdir(path = NULL, name = NULL, verbose = TRUE)
}
\arguments{
\item{path}{Drive File, something that identifies where the file should live
on your Google Drive. Can be an actual path (character), a file id marked
with \code{\link{as_id}()}, or a \code{\link{dribble}}. If specified as an
actual path, you can include the name, indicated by not appending a final
slash (i.e.: "abc/def" would indicate the file will be named "def" and placed
in "abc", where as "abc/def/" indicates the file will be placed in folder "def"
which is located in folder "abc"), or you can specify the name with the
\code{name} parameter.}

\item{name}{Character. The name of the folder you would like to create, if
not already specified in \code{path}.}

\item{verbose}{Logical, indicating whether to print informative messages
(default \code{TRUE}).}
}
\value{
An object of class \code{\link{dribble}}, a tibble with
   one row per file.
}
\description{
Create a folder on Google Drive.
}
\examples{
\dontrun{
## Create folder named "def" in existing folder "abc".
drive_mkdir(path = "abc/def")

## This will also create a folder named "def" in folder "abc".
drive_mkdir(path = "abc/", name = "def")

## Yet another way to create a folder named "def" in folder "abc",
## this time with parent folder stored in a dribble.
abc <- as_dribble("abc")
drive_mkdir(path = abc, name = "def")
}
}
