% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_mkdir.R
\name{drive_mkdir}
\alias{drive_mkdir}
\title{Create a folder on Google Drive}
\usage{
drive_mkdir(dir = NULL, path = NULL, verbose = TRUE)
}
\arguments{
\item{dir}{character, name of the folder you would like to create}

\item{path}{character, path where you would like the folder on your Google Drive}

\item{verbose}{logical, indicating whether to print informative messages
(default \code{TRUE})}
}
\value{
object of class \code{dribble} and \code{tbl_df} that contains uploaded folder's
information
}
\description{
Create a folder on Google Drive
}
