% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_mkdir.R
\name{drive_mkdir}
\alias{drive_mkdir}
\title{Make a Google Drive Directory}
\usage{
drive_mkdir(dir = NULL, path = NULL)
}
\arguments{
\item{dir}{character, name of the folder you would like to create}

\item{path}{character, path where you would like the folder on your Google Drive}
}
\value{
object of class \code{gfile} and \code{list} that contains uploaded folder's
information
}
\description{
Make a Google Drive Directory
}
