% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_mv.R
\name{drive_mv}
\alias{drive_mv}
\alias{drive_move}
\title{Move Google Drive file.}
\usage{
drive_mv(file = NULL, path = NULL, name = NULL, verbose = TRUE)

drive_move(file = NULL, path = NULL, name = NULL, verbose = TRUE)
}
\arguments{
\item{file}{Drive File, something that identifies the file of interest
on your Google Drive. This can be name(s)/path(s), marked file id(s),
or a \code{dribble}.}

\item{path}{Drive File, something that identifies where the file should live
on your Google Drive. Can be an actual path (character), a file id marked
with \code{\link{as_id}()}, or a \code{\link{dribble}}. If specified as an
actual path, you can include the name, indicated by not appending a final
slash (i.e.: "abc/def" would indicate the file will be named "def" and placed
in "abc", where as "abc/def/" indicates the file will be placed in folder "def"
which is located in folder "abc"), or you can specify the name with the
\code{name} parameter.}

\item{name}{Character, new file name if not specified as part of \code{path}. Any
name obtained from \code{path} overrides this argument. Defaults to current
name.}

\item{verbose}{Logical, indicating whether to print informative messages
(default \code{TRUE}).}
}
\value{
An object of class \code{\link{dribble}}, a tibble with
   one row per file.
}
\description{
Move a Drive file to a different folder, give it a different name, or both.
}
\examples{
\dontrun{
## create a file to move
file <- drive_upload(system.file("DESCRIPTION"), "DESC")

## rename it, but leave in current folder (root folder, in this case)
file <- drive_mv(file, "DESC-renamed")

## create a folder to move the file into
folder <- drive_mkdir("new-folder")

## move the file and rename it again
file <- drive_mv(file, folder, "DESC-re-renamed")

## verify it's in the folder
drive_ls(folder)
}
}
