% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_mv.R
\name{drive_mv}
\alias{drive_mv}
\title{Move Google Drive file}
\usage{
drive_mv(file = NULL, folder = NULL, verbose = TRUE)
}
\arguments{
\item{file}{object of class \code{dribble} or \code{drive_id} for the file you would like to move}

\item{folder}{object of class \code{dribble} or \code{drive_id} for the folder you would like to move the
file to}

\item{verbose}{logical, indicating whether to print informative messages
(default \code{TRUE})}
}
\value{
\code{dribble} that was moved
}
\description{
Move Google Drive file
}
