% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_mv.R
\name{drive_mv}
\alias{drive_mv}
\title{Move Google Drive file}
\usage{
drive_mv(file = NULL, folder = NULL, verbose = TRUE)
}
\arguments{
\item{file}{Drive File, something that identifies the file of interest
on your Google Drive. This can be name(s)/path(s), marked file id(s),
or a \code{dribble}.}

\item{folder}{Drive Folder, something that identifies the folder of interest
on your Google Drive. This can be name(s)/path(s), marked folder id(s),
or a \code{dribble}.}

\item{verbose}{logical, indicating whether to print informative messages
(default \code{TRUE})}
}
\value{
An object of class \code{dribble}, a tibble with
   one row per file.
}
\description{
Move Google Drive file
}
