% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_reveal.R
\name{drive_reveal}
\alias{drive_reveal}
\title{Add columns with pertinent information to your dribble}
\usage{
drive_reveal(file, what = "path")
}
\arguments{
\item{file}{Something that identifies the file(s) of interest on your Google
Drive. Can be a character vector of names/paths, a character vector of file
ids or URLs marked with \code{\link{as_id}()}, or a \code{\link{dribble}}.}

\item{what}{Character.
\itemize{
\item publish
\item permissions
\item trash
\item path
\item mime_type
}}
}
\value{
An object of class \code{\link{dribble}}, a tibble with one row per
  item.
}
\description{
Add columns with pertinent information to your dribble
}
\examples{
\dontrun{
## Upload a file
file <- drive_upload(R.home('doc/html/logo.jpg'))

## Reveal `trash` status
drive_reveal(file, what = "trash")

## Reveal `permissions`
drive_reveal(file, what = "permissions")

## Reveal `publish` status
drive_reveal(file, what = "publish")

## Clean up
drive_rm(file)
}
}
