% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_reveal.R
\name{drive_reveal}
\alias{drive_reveal}
\title{Add column(s) with new information}
\usage{
drive_reveal(file, what = c("path", "trashed", "mime_type", "permissions",
  "published"))
}
\arguments{
\item{file}{Something that identifies the file(s) of interest on your Google
Drive. Can be a character vector of names/paths, a character vector of file
ids or URLs marked with \code{\link{as_id}()}, or a \code{\link{dribble}}.}

\item{what}{Character, describing the type of info you want to add:
\itemize{
\item path. Warning: this can be slow, especially if called on many files.
\item trashed
\item mime_type
\item permissions. Who is this file shared with and in which roles?
\item published
}}
}
\value{
An object of class \code{\link{dribble}}, a tibble with one row per
  item.
}
\description{
Add column(s) with new information
}
\examples{
\dontrun{
## Get a nice, random selection of files
files <- drive_find(n_max = 10, q = "trashed = true or trashed = false")

## Reveal
##   * paths (warning: can be slow for many files!)
##   * if `trashed` or not
##   * MIME type
##   * permissions, i.e. sharing status
##   * if `published` or not
drive_reveal(files, "path")
drive_reveal(files, "trashed")
drive_reveal(files, "mime_type")
drive_reveal(files, "permissions")
drive_reveal(files, "published")
}
}
