% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_rm.R
\name{drive_rm}
\alias{drive_rm}
\title{Delete files from Drive.}
\usage{
drive_rm(file = NULL, verbose = TRUE)
}
\arguments{
\item{file}{Something that identifies the file(s) of interest on your Google
Drive. Can be a character vector of names/paths, a character vector of file
ids marked with \code{\link{as_id}()}, or a \code{\link{dribble}}.}

\item{verbose}{Logical, indicating whether to print informative messages
(default \code{TRUE}).}
}
\value{
Logical vector, indicating whether the delete succeeded.
}
\description{
Caution: this will permanently delete your files! For a safer,
reversible option, see \code{\link[=drive_trash]{drive_trash()}}.
}
\examples{
\dontrun{
## Create a folder to remove
folder <- drive_mkdir("folder-to-remove")

## Remove folder
drive_rm(folder)
}
}
