% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_rm.R
\name{drive_rm}
\alias{drive_rm}
\title{Delete files from Drive}
\usage{
drive_rm(..., verbose = TRUE)
}
\arguments{
\item{verbose}{Logical, indicating whether to print informative messages
(default \code{TRUE}).}

\item{file}{Something that identifies the file(s) of interest on your Google
Drive. Can be a character vector of names/paths, a character vector of file
ids or URLs marked with \code{\link{as_id}()}, or a \code{\link{dribble}}.}
}
\value{
Logical vector, indicating whether the delete succeeded.
}
\description{
Caution: this will permanently delete your files! For a safer, reversible
option, see \code{\link[=drive_trash]{drive_trash()}}.
}
\examples{
\dontrun{
## Create something to remove
drive_upload(system.file("DESCRIPTION"), name = "abc-DESC")

## Remove it by name
drive_rm("abc-DESC")

## Create several things to remove
x1 <- drive_upload(system.file("DESCRIPTION"), name = "abc-DESC")
drive_upload(system.file("DESCRIPTION"), name = "def-DESC")
x2 <- drive_upload(system.file("DESCRIPTION"), name = "ghi-DESC")

## Remove them all at once, specified in different ways
drive_rm(x1, "def-DESC", as_id(x2))
}
}
\seealso{
Wraps the \code{files.delete} endpoint:
\itemize{
\item \url{https://developers.google.com/drive/v3/reference/files/delete}
}
}
