% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_search.R
\name{drive_search}
\alias{drive_search}
\title{Search for files on Google Drive.}
\usage{
drive_search(pattern = NULL, type = NULL, ..., verbose = TRUE)
}
\arguments{
\item{pattern}{Character. If provided, only the files whose names match this
regular expression are returned.}

\item{type}{Character. If provided, only files of this type will be returned.
Can be anything that \code{\link[=drive_mime_type]{drive_mime_type()}} knows how to handle. This is
processed by googledrive and sent as a query parameter.}

\item{...}{Query parameters to pass along to the API query.}

\item{verbose}{Logical, indicating whether to print informative messages
(default \code{TRUE}).}
}
\value{
An object of class \code{\link{dribble}}, a tibble with
   one row per file.
}
\description{
This is the closest googledrive function to what you get from
\url{https://drive.google.com}: by default, you just get a listing of your files.
You can also narrow the search in various ways, such as by file type, whether
it's yours or shared with you, starred status, etc.
}
\examples{
\dontrun{
## list "My Drive" w/o regard for folder hierarchy
drive_search()

## search for files located directly in your root folder
drive_search(q = "'root' in parents")

## filter for folders
drive_search(type = "folder")
drive_search(q = "mimeType = 'application/vnd.google-apps.folder'")

## filter for Google Sheets
drive_search(type = "spreadsheet")
drive_search(q = "mimeType='application/vnd.google-apps.spreadsheet'")

## files whose names match a regex
drive_search(pattern = "jt")

## stuff about pagination
drive_search(pageSize = 50)
}

}
\seealso{
Helpful links for forming queries: *
\url{https://developers.google.com/drive/v3/web/search-parameters} *
\url{https://developers.google.com/drive/v3/reference/files/list}
}
