% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_share.R
\name{drive_share}
\alias{drive_share}
\title{Update Google Drive file share permissions.}
\usage{
drive_share(file = NULL, role = NULL, type = NULL, ..., verbose = TRUE)
}
\arguments{
\item{file}{Drive File, something that identifies the file of interest
on your Google Drive. This can be name(s)/path(s), marked file id(s),
or a \code{dribble}.}

\item{role}{The role granted by this permission. Valid values are:
\itemize{
\item organizer
\item owner
\item writer
\item commenter
\item reader
}}

\item{type}{The type of the grantee. Valid values are:
\itemize{
\item user
\item group
\item domain
\item anyone
}}

\item{...}{Name-value pairs to add to the API request body.}

\item{verbose}{Logical, indicating whether to print informative messages
(default \code{TRUE}).}
}
\value{
An object of class \code{\link{dribble}}, a tibble with
   one row per file.
}
\description{
Update Google Drive file share permissions.
}
