% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_share_link.R
\name{drive_share_link}
\alias{drive_share_link}
\title{Retrieve a file's share link.}
\usage{
drive_share_link(file)
}
\arguments{
\item{file}{Something that identifies the file(s) of interest on your Google
Drive. Can be a character vector of names/paths, a character vector of file
ids marked with \code{\link{as_id}()}, or a \code{\link{dribble}}.}
}
\value{
Character. The link to Google Drive file.
}
\description{
Retrieve a file's share link.
}
