% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_share_link.R
\name{drive_share_link}
\alias{drive_share_link}
\title{Retrieves Google Drive file's share link}
\usage{
drive_share_link(file)
}
\arguments{
\item{file}{\code{dribble} representing the file you would like to
retrieve the link for}
}
\value{
character, link to Google Drive file
}
\description{
Retrieves Google Drive file's share link
}
