% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_auth.R
\name{drive_token}
\alias{drive_token}
\title{Produce Google token}
\usage{
drive_token(verbose = FALSE)
}
\arguments{
\item{verbose}{Logical, indicating whether to print informative messages
(default \code{TRUE}).}
}
\value{
a \code{request} object (an S3 class provided by \link[httr:httr]{httr})
}
\description{
If token is not already available, call \code{\link[=drive_auth]{drive_auth()}} to either load from
cache or initiate OAuth2.0 flow. Return the token -- not "bare" but, rather,
prepared for inclusion in downstream requests. Use \code{access_token()} to reveal
the actual access token, suitable for use with curl.
}
\keyword{internal}
