% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_auth.R
\name{drive_token}
\alias{drive_token}
\alias{include_token_if}
\alias{omit_token_if}
\title{Produce Google token}
\usage{
drive_token(verbose = FALSE)

include_token_if(cond)

omit_token_if(cond)
}
\arguments{
\item{verbose}{logical, indicating whether to print informative messages
(default \code{TRUE})}
}
\value{
a \code{request} object (an S3 class provided by \code{httr})
}
\description{
If token is not already available, call \code{[drive_auth()]} to either load
from cache or initiate OAuth2.0 flow. Return the token -- not "bare" but,
rather, prepared for inclusion in downstream requests. Use
\code{access_token()} to reveal the actual access token, suitable for use
with \code{curl}.
}
\keyword{internal}
