% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_trash.R
\name{drive_trash}
\alias{drive_trash}
\alias{drive_untrash}
\title{Move Drive files to the trash.}
\usage{
drive_trash(file = NULL, verbose = TRUE)

drive_untrash(file = NULL, verbose = TRUE)
}
\arguments{
\item{file}{Drive File, something that identifies the file of interest
on your Google Drive. This can be name(s)/path(s), marked file id(s),
or a \code{dribble}.}

\item{verbose}{Logical, indicating whether to print informative messages
(default \code{TRUE}).}
}
\value{
Logical vector, indicating whether the trashing/untrashing succeeded.
}
\description{
Move Drive files to the trash.
}
\examples{
\dontrun{
## Create a file to trash.
file <- drive_upload(system.file("DESCRIPTION"), "DESC")
## Put a single file in the trash.
drive_trash("DESC")
drive_untrash("DESC")

## Put multiple files in trash.
file_a <- drive_upload(system.file("DESCRIPTION"), "file_a")
file_b <- drive_upload(system.file("DESCRIPTION"), "file_b")
drive_trash(c("file_a", "file_b"))
drive_untrash(c("file_a", "file_b"))
}
}
