% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_publish.R
\name{drive_unpublish}
\alias{drive_unpublish}
\title{Unpublish Google Drive file}
\usage{
drive_unpublish(file, ..., verbose = TRUE)
}
\arguments{
\item{file}{Something that identifies the file(s) of interest on your Google
Drive. Can be a character vector of names/paths, a character vector of file
ids or URLs marked with \code{\link{as_id}()}, or a \code{\link{dribble}}.}

\item{...}{Name-value pairs to add to the API request body.}

\item{verbose}{Logical, indicating whether to print informative messages
(default \code{TRUE}).}
}
\value{
An object of class \code{\link{dribble}}, a tibble with one row per
  item.
}
\description{
Unpublish a Google Drive file to the web. You can only publish "Google" files, such
as Google Documents, Google Spreadsheets, or Google Slides, therefore you cannot
(and there would be no need to) unpublish non-"Google" type files.
}
\examples{
\dontrun{
#' ## Upload file to publish
file <- drive_upload(R.home('doc/BioC_mirrors.csv'),
  type = "spreadsheet")

## Publish file
file <- drive_publish(file)

## Unplublish file
file <- drive_unpublish(file)

## Clean up
drive_rm(file)
}
}
\seealso{
Wraps the \code{revisions.update} endpoint:
\itemize{
\item \url{https://developers.google.com/drive/v3/reference/revisions/update}
}
}
