% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_publish.R
\name{drive_unpublish}
\alias{drive_unpublish}
\title{Unpublish Google Drive file}
\usage{
drive_unpublish(file = NULL, ..., verbose = TRUE)
}
\arguments{
\item{file}{\code{dribble} or \code{drive_id} object representing the file you would
like to unpublish}

\item{...}{name-value pairs to add to the API request body.}

\item{verbose}{logical, indicating whether to print informative messages
(default \code{TRUE})}
}
\value{
\code{dribble} object, with published information as a \code{tibble}
added under the column name \code{publish}
}
\description{
Unpublish Google Drive file
}
