% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_update.R
\name{drive_update}
\alias{drive_update}
\title{Update a file on Drive with new content.}
\usage{
drive_update(file = NULL, path = NULL, verbose = TRUE)
}
\arguments{
\item{file}{Character, path to the local file to upload.}

\item{path}{Specifies target destination for \code{file} or copy of
\code{file} on Google Drive. Can be an actual path (character), a file id
marked with \code{\link{as_id}()}, or a \code{\link{dribble}}. If specified
as an actual path, it is best to explicitly indicate if it's a folder by
including a trailing slash, since it cannot always be worked out from the
context of the call.}

\item{verbose}{Logical, indicating whether to print informative messages
(default \code{TRUE}).}
}
\value{
An object of class \code{\link{dribble}}, a tibble with
   one row per file.
}
\description{
Update a file on Drive with new content.
}
\examples{
\dontrun{
## Upload a file to update.
x <- drive_upload(R.home("doc/NEWS"))

## Update the file.
x <- drive_update(R.home("doc/NEWS.2"), x)

## clean-up
drive_rm(x)
}
}
