% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_upload.R
\name{drive_upload}
\alias{drive_upload}
\title{Upload a file to Google Drive.}
\usage{
drive_upload(from = NULL, path = NULL, name = NULL, overwrite = FALSE,
  type = NULL, verbose = TRUE)
}
\arguments{
\item{from}{Character, local path to the file to upload.}

\item{path}{Drive File, something that identifies where the file should live
on your Google Drive. Can be an actual path (character), a file id marked
with \code{\link{as_id}()}, or a \code{\link{dribble}}. If specified as an
actual path, you can include the name, indicated by not appending a final
slash (i.e.: "abc/def" would indicate the file will be named "def" and placed
in "abc", where as "abc/def/" indicates the file will be placed in folder "def"
which is located in folder "abc"), or you can specify the name with the
\code{name} parameter.}

\item{name}{Character, name the file should have on Google Drive if not specified
in \code{path}. Will default to its local name.}

\item{overwrite}{A logical scalar, do you want to overwrite a file already on
Google Drive, if such exists?}

\item{type}{Character. If type = \code{NULL}, a MIME type is automatically
determined from the file extension, if possible. If the source file is of a
suitable type, you can request conversion to Google Doc, Sheet or Slides by
setting \code{type} to \code{document}, \code{spreadsheet}, or \code{presentation},
respectively. All non-\code{NULL} values for \code{type} are pre-processed with
\code{\link[=drive_mime_type]{drive_mime_type()}}.}

\item{verbose}{Logical, indicating whether to print informative messages
(default \code{TRUE}).}
}
\value{
An object of class \code{\link{dribble}}, a tibble with
   one row per file.
}
\description{
Upload a file to Google Drive.
}
\examples{
\dontrun{
write.csv(chickwts, "chickwts.csv")
drive_chickwts <- drive_upload("chickwts.csv")

## or convert it to a Google Sheet
drive_chickwts <- drive_upload("chickwts.csv", type = "spreadsheet")
}
}
\seealso{
MIME types that can be converted to native Google formats:
\itemize{
\item \url{https://developers.google.com/drive/v3/web/manage-uploads#importing_to_google_docs_types_wzxhzdk18wzxhzdk19}
}
}
