% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_upload.R
\name{drive_upload}
\alias{drive_upload}
\title{Upload a file to Google Drive}
\usage{
drive_upload(from = NULL, up_name = NULL, up_folder = NULL,
  overwrite = FALSE, type = NULL, verbose = TRUE)
}
\arguments{
\item{from}{character, local path to the file to upload}

\item{up_name}{character, name the file should have on Google Drive. Will
default to its local name.}

\item{up_folder}{character, name of parent folder on Google Drive. Will
default to user's root folder, i.e. the top-level of "My Drive".}

\item{overwrite}{logical, do you want to overwrite a file already on Google
Drive, if such exists?}

\item{type}{character. If type = \code{NULL}, a MIME type is automatically
determined from the file extension, if possible. If the source file is of a
suitable type, you can request conversion to Google Doc, Sheet or Slides by
setting \code{type} to \code{document}, \code{spreadsheet}, or \code{presentation},
respectively.}

\item{verbose}{logical, indicating whether to print informative messages
(default \code{TRUE})

MIME types that can be converted to native Google formats:
\itemize{
\item \url{https://developers.google.com/drive/v3/web/manage-uploads#importing_to_google_docs_types_wzxhzdk18wzxhzdk19}
}}
}
\value{
object of class \code{gfile} and \code{list} that contains uploaded file's
information
}
\description{
Upload a file to Google Drive
}
