% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_upload.R
\name{drive_upload}
\alias{drive_upload}
\title{Upload a file to Google Drive}
\usage{
drive_upload(from = NULL, up_name = NULL, up_folder = NULL,
  overwrite = FALSE, type = NULL, verbose = TRUE)
}
\arguments{
\item{from}{character, local path to the file to upload}

\item{up_name}{character, name the file should have on Google Drive. Will
default to its local name.}

\item{up_folder}{character, name of parent folder on Google Drive. Will
default to user's root folder, i.e. the top-level of "My Drive".}

\item{overwrite}{logical, do you want to overwrite a file already on Google
Drive, if such exists?}

\item{type}{character. If type = \code{NULL}, a MIME type is automatically
determined from the file extension, if possible. If the source file is of a
suitable type, you can request conversion to Google Doc, Sheet or Slides by
setting \code{type} to \code{document}, \code{spreadsheet}, or \code{presentation},
respectively.}

\item{verbose}{logical, indicating whether to print informative messages
(default `TRUE`)}
}
\value{
Drive File. An object of class `dribble`, a tibble with
   one row per file.
}
\description{
Upload a file to Google Drive
}
\examples{
MIME types that can be converted to native Google formats:
   * <https://developers.google.com/drive/v3/web/manage-uploads#importing_to_google_docs_types_wzxhzdk18wzxhzdk19>

\dontrun{
write.csv(chickwts, "chickwts.csv")
drive_chickwts <- drive_upload("chickwts.csv")

## or convert it to a Google Sheet
drive_chickwts <- drive_upload("chickwts.csv", type = "spreadsheet")
}
}
