% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_upload.R
\name{drive_upload}
\alias{drive_upload}
\title{Upload a file to Google Drive}
\usage{
drive_upload(from = NULL, up_name = NULL, up_folder = NULL,
  overwrite = FALSE, type = NULL, verbose = TRUE)
}
\arguments{
\item{from}{character, local path to the file to upload}

\item{up_name}{character, name the file should have on Google Drive. Will
default to its local name.}

\item{up_folder}{character, name of parent folder on Google Drive. Will
default to user's root folder, i.e. the top-level of "My Drive".}

\item{overwrite}{logical, do you want to overwrite file already on Google
Drive}

\item{type}{if type = \code{NULL}, will upload as a non-Google Drive document
otherwise you can specify \code{document}, \code{spreadsheet}, or \code{presentation}.
Files with no extension will be assumed to be a \code{folder}.}

\item{verbose}{logical, indicating whether to print informative messages
(default \code{TRUE})}
}
\value{
object of class \code{gfile} and \code{list} that contains uploaded file's
information
}
\description{
Upload a file to Google Drive
}
