% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_upload.R
\name{drive_upload}
\alias{drive_upload}
\title{Upload a file to Google Drive}
\usage{
drive_upload(input = NULL, output = NULL, overwrite = FALSE,
  type = NULL, ..., verbose = TRUE)
}
\arguments{
\item{input}{character, path (on your computer) of the file you'd like to upload}

\item{output}{character, path (on your Google Drive) where you'd like to upload
the file}

\item{overwrite}{logical, do you want to overwrite file already on Google
Drive}

\item{type}{if type = \code{NULL}, will upload as a non-Google Drive document
otherwise you can specify \code{document}, \code{spreadsheet}, or \code{presentation}.
Files with no extension will be assumed to be a \code{folder}.}

\item{...}{name-value pairs to query the API}

\item{verbose}{logical, indicating whether to print informative messages
(default \code{TRUE})}
}
\value{
object of class \code{gfile} and \code{list} that contains uploaded file's
information
}
\description{
Upload a file to Google Drive
}
