% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_upload.R
\name{drive_upload}
\alias{drive_upload}
\title{Upload a file to Drive.}
\usage{
drive_upload(file = NULL, path = NULL, name = NULL, overwrite = FALSE,
  type = NULL, verbose = TRUE)
}
\arguments{
\item{file}{Character, path to the local file to upload.}

\item{path}{Specifies target destination for \code{file} or copy of
\code{file} on Google Drive. Can be an actual path (character), a file id
marked with \code{\link{as_id}()}, or a \code{\link{dribble}}. If specified
as an actual path, it is best to explicitly indicate if it's a folder by
including a trailing slash, since it cannot always be worked out from the
context of the call.}

\item{name}{Character, name the file should have on Google Drive if not
specified in \code{path}. Will default to its local name.}

\item{overwrite}{A logical scalar, do you want to overwrite a file already on
Google Drive, if such exists? If you intend to overwrite an existing
\code{dribble} by specifying it using \code{path} and \code{overwrite = TRUE}, leave
\code{name} \code{NULL}. You can rename this file using \code{drive_rename()}.}

\item{type}{Character. If type = \code{NULL}, a MIME type is automatically
determined from the file extension, if possible. If the source file is of a
suitable type, you can request conversion to Google Doc, Sheet or Slides by
setting \code{type} to \code{document}, \code{spreadsheet}, or \code{presentation},
respectively. All non-\code{NULL} values for \code{type} are pre-processed with
\code{\link[=drive_mime_type]{drive_mime_type()}}.}

\item{verbose}{Logical, indicating whether to print informative messages
(default \code{TRUE}).}
}
\value{
An object of class \code{\link{dribble}}, a tibble with
   one row per file.
}
\description{
Upload a file to Drive.
}
\examples{
\dontrun{
## upload a csv file
mirrors_csv <- drive_upload(R.home('doc/BioC_mirrors.csv'))

## or convert it to a Google Sheet
mirrors_sheet <- drive_upload(
  R.home('doc/BioC_mirrors.csv'),
  name = "BioC_mirrors",
  type = "spreadsheet"
)

## check out the new Sheet!
drive_browse(mirrors_sheet)

## clean-up
drive_find("BioC_mirrors") \%>\% drive_rm()
}
}
\seealso{
MIME types that can be converted to native Google formats:
\itemize{
\item \url{https://developers.google.com/drive/v3/web/manage-uploads#importing_to_google_docs_types_wzxhzdk18wzxhzdk19}
}
}
