% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_upload.R
\name{drive_upload}
\alias{drive_upload}
\title{Upload into a new Drive file}
\usage{
drive_upload(media, path = NULL, name = NULL, type = NULL, ...,
  verbose = TRUE)
}
\arguments{
\item{media}{Character, path to the local file to upload.}

\item{path}{Specifies target destination for the new file on Google
Drive. Can be an actual path (character), a file id marked with
\code{\link{as_id}()}, or a \code{\link{dribble}}. If specified as an
actual path, it is best to explicitly indicate if it's a folder by
including a trailing slash, since it cannot always be worked out from the
context of the call. Will default to its local name.}

\item{name}{Character, new file name if not specified as part of
\code{path}. This will force \code{path} to be treated as a folder, even if
it is character and lacks a trailing slash. Will default to its local name.}

\item{type}{Character. If \code{type = NULL}, a MIME type is automatically
determined from the file extension, if possible. If the source file is of a
suitable type, you can request conversion to Google Doc, Sheet or Slides by
setting \code{type} to \code{document}, \code{spreadsheet}, or \code{presentation},
respectively. All non-\code{NULL} values for \code{type} are pre-processed with
\code{\link[=drive_mime_type]{drive_mime_type()}}.}

\item{...}{Named parameters to pass along to the Drive API. You can affect
the metadata of the target file by specifying properties of the Files
resource via \code{...}. See the "Request body" section of the Drive API
docs for the associated endpoint.}

\item{verbose}{Logical, indicating whether to print informative messages
(default \code{TRUE}).}
}
\value{
An object of class \code{\link{dribble}}, a tibble with one row per
  item.
}
\description{
Uploads a local file into a new Drive file. To update the content or metadata
of an existing Drive file, use \code{\link[=drive_update]{drive_update()}}.
}
\examples{
\dontrun{
## upload a csv file
mirrors_csv <- drive_upload(R.home('doc/BioC_mirrors.csv'))

## or convert it to a Google Sheet
mirrors_sheet <- drive_upload(
  R.home('doc/BioC_mirrors.csv'),
  name = "BioC_mirrors",
  type = "spreadsheet"
)

## check out the new Sheet!
drive_browse(mirrors_sheet)

## clean-up
drive_find("BioC_mirrors") \%>\% drive_rm()

## Upload a file and, at the same time, star it
logo <- drive_upload(
  R.home('doc/html/logo.jpg'),
  starred = "true"
)

## Clean up
drive_rm(logo)
}
}
\seealso{
Wraps the \code{files.create} endpoint:
\itemize{
\item \url{https://developers.google.com/drive/v3/reference/files/create}
}

MIME types that can be converted to native Google formats:
\itemize{
\item \url{https://developers.google.com/drive/v3/web/manage-uploads#importing_to_google_docs_types_wzxhzdk18wzxhzdk19}
}
}
