% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_upload.R
\name{drive_upload}
\alias{drive_upload}
\title{Upload a file to Google Drive.}
\usage{
drive_upload(from = NULL, name = NULL, folder = NULL, overwrite = FALSE,
  type = NULL, verbose = TRUE)
}
\arguments{
\item{from}{Character, local path to the file to upload.}

\item{name}{Character, name the file should have on Google Drive. Will
default to its local name.}

\item{folder}{Drive File, something that identifies the folder of interest
on your Google Drive. This can be name(s)/path(s), marked file id(s),
or a \code{dribble}.}

\item{overwrite}{A logical scalar, do you want to overwrite a file already on
Google Drive, if such exists?}

\item{type}{Character. If type = \code{NULL}, a MIME type is automatically
determined from the file extension, if possible. If the source file is of a
suitable type, you can request conversion to Google Doc, Sheet or Slides by
setting \code{type} to \code{document}, \code{spreadsheet}, or \code{presentation},
respectively. All non-\code{NULL} values for `type`` are pre-processed with
\code{\link[=drive_mime_type]{drive_mime_type()}}.}

\item{verbose}{Logical, indicating whether to print informative messages
(default \code{TRUE}).}
}
\value{
An object of class \code{\link{dribble}}, a tibble with
   one row per file.
}
\description{
Upload a file to Google Drive.
}
\examples{
\dontrun{
write.csv(chickwts, "chickwts.csv")
drive_chickwts <- drive_upload("chickwts.csv")

## or convert it to a Google Sheet
drive_chickwts <- drive_upload("chickwts.csv", type = "spreadsheet")
}
}
\seealso{
MIME types that can be converted to native Google formats:
\itemize{
\item \url{https://developers.google.com/drive/v3/web/manage-uploads#importing_to_google_docs_types_wzxhzdk18wzxhzdk19}
}
}
