% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_user.R
\name{drive_user}
\alias{drive_user}
\title{Google Drive User Information}
\usage{
drive_user(fields = "user", ...)
}
\arguments{
\item{fields}{fields to query, default is \code{user}.}

\item{...}{name-value pairs to query the API}
}
\value{
list of class \code{drive_user} with user's information
}
\description{
Google Drive User Information
}
\keyword{internal}
