% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_user.R
\name{drive_user}
\alias{drive_user}
\title{Get info on current user}
\usage{
drive_user(verbose = deprecated())
}
\arguments{
\item{verbose}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} This logical argument to
individual googledrive functions is deprecated. To globally suppress
googledrive messaging, use \code{options(googledrive_quiet = TRUE)} (the default
behaviour is to emit informational messages). To suppress messaging in a
more limited way, use the helpers \code{\link[=local_drive_quiet]{local_drive_quiet()}} or
\code{\link[=with_drive_quiet]{with_drive_quiet()}}.}
}
\value{
A list of class \code{drive_user}.
}
\description{
Reveals information about the user associated with the current token. This is
a thin wrapper around \code{\link[=drive_about]{drive_about()}} that just extracts the most useful
information (the information on current user) and prints it nicely.
}
\examples{
\dontshow{if (drive_has_token()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
drive_user()

# more info is returned than is printed
user <- drive_user()
str(user)
\dontshow{\}) # examplesIf}
}
\seealso{
Wraps the \code{about.get} endpoint:
\itemize{
\item \url{https://developers.google.com/drive/v3/reference/about/get}
}
}
