% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_publish.R
\name{gd_check_publish}
\alias{gd_check_publish}
\title{Check if Google Drive file is published}
\usage{
gd_check_publish(file, verbose = TRUE)
}
\arguments{
\item{file}{\code{drive_file} object representing the file you would like to check the published status of}

\item{verbose}{logical, indicating whether to print informative messages (default \code{TRUE})}
}
\value{
logical, whether the file is published
}
\description{
Check if Google Drive file is published
}
