% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_ls.R
\name{gd_ls}
\alias{gd_ls}
\title{Tibble of files on Google Drive}
\usage{
gd_ls(search = NULL, ..., fixed = FALSE, verbose = TRUE)
}
\arguments{
\item{search}{character, regular expression(s) of title(s) of documents to output in a tibble.
If it is \code{NULL} (default), information about all documents in drive will be output in a tibble.}

\item{...}{name-value pairs to query the API}

\item{fixed}{logical, from \code{[grep()]}. If \code{TRUE``,}search` is exactly matched to a
document's name on Google Drive.}

\item{verbose}{logical, indicating whether to print informative messages (default \code{TRUE})

This will default to the most recent 100 files on your Google Drive. If you would like more than 100,
include using the \code{pageSize} parameter. For example, if I wanted 200, I would run \code{gd_ls(pageSize = 200)}.}
}
\value{
tibble containing the name, type, and id of files on your google drive (default 100 files)
}
\description{
Tibble of files on Google Drive
}
