% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{get_file}
\alias{get_file}
\title{Download non-Google docs resources from Google Drive}
\usage{
get_file(fileID)
}
\arguments{
\item{fileID}{ID of the file in Google Drive}
}
\description{
The function will acknowledge that the file is safe to download. Do be careful when
downloading files from the web (even if its on your own Google Drive folder)
}

