% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{googledrive-deprecated}
\alias{googledrive-deprecated}
\alias{drive_auth_config}
\title{Deprecated googledrive functions}
\usage{
drive_auth_config(app, path, api_key)
}
\arguments{
\item{app}{OAuth app, in the sense of \code{\link[httr:oauth_app]{httr::oauth_app()}}.}

\item{path}{JSON downloaded from Google Cloud Platform Console, containing a
client id (aka key) and secret, in one of the forms supported for the \code{txt}
argument of \code{\link[jsonlite:fromJSON]{jsonlite::fromJSON()}} (typically, a file path or JSON string).}

\item{api_key}{API key.}
}
\description{
Deprecated googledrive functions
}
\keyword{internal}
