% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googledrive-package.R
\docType{package}
\name{googledrive-package}
\alias{googledrive}
\alias{googledrive-package}
\title{googledrive: An Interface to Google Drive}
\description{
TO DO: mine index.Rmd and articles for some material
}
\seealso{
Useful links:
\itemize{
  \item \url{http://googledrive.tidyverse.org}
  \item \url{https://github.com/tidyverse/googledrive}
  \item Report bugs at \url{https://github.com/tidyverse/googledrive/issues}
}

}
\author{
\strong{Maintainer}: Jennifer Bryan \email{jenny@rstudio.com}

Authors:
\itemize{
  \item Lucy D'Agostino McGowan \email{ld.mcgowan@vanderbilt.edu}
}

Other contributors:
\itemize{
  \item RStudio [copyright holder, funder]
}

}
\keyword{internal}
