% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dribble.R
\name{is_one}
\alias{is_one}
\title{Require that dribble contains exactly one Drive file.}
\usage{
is_one(d, .what = "file")
}
\arguments{
\item{d}{A \code{\link{dribble}}.}

\item{.what}{Character, description of input for informative
error message.}
}
\description{
This will return the input \code{\link{dribble}} if it contains exactly
one Drive file, and will error otherwise.
}
