% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_request.R
\name{make_request}
\alias{make_request}
\title{Make a request for the Google Drive v3 API}
\usage{
make_request(x, ...)
}
\arguments{
\item{x}{List, contains the  \code{method}, \code{path}, \code{query}, and \code{url}, to make
the API request obtained from \code{\link[=build_request]{build_request()}}}

\item{...}{List, Name-value pairs to query the API}
}
\value{
Object of class \code{response} from \link{httr}.
}
\description{
Make a request for the Google Drive v3 API
}
