% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{update_file_metadata}
\alias{update_file_metadata}
\title{Update file metadata properties}
\usage{
update_file_metadata(fileId, name = NULL, description = NULL,
  starred = NULL, trashed = NULL)
}
\arguments{
\item{fileId}{The ID of the file.}

\item{name}{Name of the file. Overwrites the file name on Google Drive}

\item{description}{Description of the file. Overwrites the description on Google Drive}
}
\description{
Allows you to update the name and description of the file
}

