% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{upload_file}
\alias{upload_file}
\title{Upload file to Google Drive}
\usage{
upload_file(fileName)
}
\description{
Allows you to uploads files into Google Drive. During the uploading process, you would
not be able to define other metadata that concerns the file. Utilize other functions to edit the
file metadata
}

